==> R_Skripte/13_Deskription_Gruppen.Rmd <==
---
title: "Deskription Gruppen"
author: "ldp"
date: "2022/08/16"
output:
  html_document:
    toc: true
    toc_float:
      collapsed: false
---

```{r setup, include=FALSE}
library(tidyverse) 
library(lubridate)
library(psych)
library(dlookr)
library(table1)
library(janitor)
library(openxlsx)
#library(epiDisplay)
```


# 2019

```{r data_load_2019}
setwd("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/02_rdata/2019")

load("data11_neu.Rda")
load("kh11.Rda")

data1 <- data11_neu
```


```{r kh_prep_2019, include = FALSE}
### Data prep
# Dauer des Intervalls
data1<- data1 %>% 
  mutate(intervalldauer = as.double(END_DATE - START_DATE +1),
         tod = if_else(is.na(V_STERBEDATUM),0,1))

# KH-Daten vorbereiten
# nur vollstationär behalten
kh11<- kh11  %>%
  mutate(KH_AUF_ANLASS2 = str_pad(KH_AUF_ANLASS, 4, pad = "0"),
         KH_AUF_ANLASS_Beg = substr(KH_AUF_ANLASS2, 0, 2),
         KH_AUF_ANLASS_End = substr(KH_AUF_ANLASS2, 3, 4)) %>% 
  filter(KH_AUF_ANLASS_Beg=="01")

# überschneidende Hospitalisierungen verbinden
kh11$KH_BEGINN_DAT <- ymd(kh11$KH_BEGINN_DAT)
kh11$KH_ENDE_DAT <- ymd(kh11$KH_ENDE_DAT)

kh11a <- kh11 %>%
  arrange(V_ID, KH_BEGINN_DAT) %>% 
  group_by(V_ID) %>%
  mutate(indx = c(0, cumsum(as.numeric(lead(KH_BEGINN_DAT )) > cummax(as.numeric(KH_ENDE_DAT )))[-n()])) %>%
  group_by(V_ID, indx) %>%
  summarise(KH_BEGINN_DAT2 = min(KH_BEGINN_DAT), 
            KH_ENDE_DAT2 = max(KH_ENDE_DAT))

# Kosten und Uhrzeit des ersten Falles betrachten 
kh11b <- left_join(kh11a, kh11, by=c("V_ID", "KH_BEGINN_DAT2"="KH_BEGINN_DAT"))

# bei mehreren Fällen mit Beginn am Aufnahmetag, Fall mit längster Aufenthaltsdauer behalten
kh11c <- kh11b %>% 
  group_by(V_ID, KH_BEGINN_DAT2) %>% 
  arrange(desc(KH_ENDE_DAT)) %>% 
  mutate(n = paste0( row_number())) %>% 
  filter(n==1)

# Stammdaten mit Krankenhausdaten verbinden
data3 <- left_join(data1, kh11c, by=("V_ID"))

data3a <- data3 %>% 
  filter(!(KH_BEGINN_DAT2 >= END_DATE | KH_ENDE_DAT2 <= START_DATE)) %>% 
  rowwise() %>% 
  mutate(rel_KH_BEGINN_DAT = max(KH_BEGINN_DAT2,START_DATE),
         rel_KH_ENDE_DAT = min(KH_ENDE_DAT2,END_DATE)) %>% 
  dplyr::mutate(kh_dauer = as.double(rel_KH_ENDE_DAT - rel_KH_BEGINN_DAT + 1))

# Faktorvariablen für group_by (drop false) notwendig
data1 <- data1 %>% mutate(tg = as.factor(tg), 
                          V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD), 
                          V_ALTER = as.factor(V_ALTER)) 

data3a <- data3a %>% 
  mutate(tg = as.factor(tg), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD), 
         V_ALTER = as.factor(V_ALTER)) 
```



```{r verweildauer_2019}
verweildauer_zg3 <- data3a %>% 
  filter(ZG3==1) %>% 
  group_by(V_ID) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4 <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_TG  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,TG) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_TG <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,TG) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_pg  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_pg <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_pg_TG  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_pg_TG <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_pg_TG_tod <- data3a %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD, V_ALTER,tod) %>% 
  summarise(Verweildauer = sum(kh_dauer))

list_of_datasets <- list("verweildauer_zg3" = verweildauer_zg3,
                         "verweildauer_zg4" = verweildauer_zg4,
                         "verweildauer_zg3_TG" = verweildauer_zg3_TG, 
                         "verweildauer_zg4_TG" = verweildauer_zg4_TG,
                         "verweildauer_zg3_pg" = verweildauer_zg3_pg, 
                         "verweildauer_zg4_pg" = verweildauer_zg4_pg,
                         "verweildauer_zg3_pg_TG" = verweildauer_zg3_pg_TG, 
                         "verweildauer_zg4_pg_TG" = verweildauer_zg4_pg_TG,
                         "verweildauer_pg_TG_tod" =verweildauer_pg_TG_tod)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2019.xlsx")
```
```{r descr_alle_2019}
# Dauer des Intervalls
data1<- data1 %>% 
  mutate(intervalldauer = as.double(END_DATE - START_DATE +1),
         tod = if_else(is.na(V_STERBEDATUM),0,1))

# ZG3 
tab1_zg3_bas <- data1 %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer)) 

# ZG3-KH
tab1_zg3_kh <- data3a %>%
  ungroup() %>% 
  summarise(KH_tage= sum(kh_dauer)) 

# Daten verbinden
tab1_zg3 <- bind_cols(tab1_zg3_bas, tab1_zg3_kh)

tab1_alle <- tab1_zg3 %>% 
  mutate(oKH_tage= VZ_tage-KH_tage,
         VZ_tage_N = VZ_tage/N,
         KH_tage_N = KH_tage/N,
         oKH_tage_N = oKH_tage/N,
         per_KH_Tage= KH_tage/VZ_tage)

data1a <- data1 %>% 
  group_by(V_ID) %>% 
  arrange(START_DATE) %>% 
  mutate(n=row_number())

# nur erstes Intervall pro ID behalten
table(data1a$n)
data1b <- data1a %>% filter(n==1)

data1b$V_ALTER<- 
  factor(data1b$V_ALTER , 
         levels = c(3,4,5,6,7,8,9),
         labels = c("60-64","65-69","70-74","75-79","80-84","85-89","90 und mehr" ))

label(data1b$V_ALTER) <- "Altersgruppe"

data1b$tod<- 
  factor(data1b$tod , 
         levels=c(0,1),
         labels=c("nein", "ja"))

label(data1b$tod) <- "Tod"

label(data1b$V_PFLEGEGRAD) <- "Pflegegrad"

table1(~ V_ALTER + tod + V_PFLEGEGRAD| TG, data=data1b)
table1 <- table1(~ V_ALTER + tod + V_PFLEGEGRAD| TG, data=data1b)
table1 <- as.data.frame(table1)
```


```{r descr2_alle_2019, include = FALSE}
# Ausschluss Daten vor 2019-02-01
data1e <- data1 %>% filter(END_DATE>"2019-01-31")
data1e$START_DATE [data1e$START_DATE<"2019-02-01"] <-"2019-02-01" 

data1e<- data1e %>% 
  dplyr::mutate(intervalldauer = as.double(END_DATE - START_DATE +1))

data1f<- data1e %>% group_by(V_ID) %>% 
  arrange(START_DATE) %>% 
  mutate(n=row_number())

# nur erstes Intervall pro ID behalten
table(data1f$n)
data1g <- data1f %>% filter(n==1)

data1g$V_ALTER<- 
  factor(data1g$V_ALTER , 
         levels=c(3,4,5,6,7,8,9),
         labels=c("60-64","65-69","70-74","75-79","80-84","85-89","90 und mehr" ))

label(data1g$V_ALTER) <- "Altersgruppe"

data1g$tod <- 
  factor(data1g$tod , 
         levels=c(0,1),
         labels=c("nein", "ja"))

label(data1g$tod) <- "Tod"
label(data1g$V_PFLEGEGRAD) <- "Pflegegrad"

table1_feb <-table1(~ V_ALTER + tod+ V_PFLEGEGRAD | TG, data=data1g)
table1_feb <- as.data.frame(table1_feb)



### Pflegegrad am 30.06.2019 (ehemals 31.12.2020)
data1c <- data1 %>% 
  filter(END_DATE >= "2019-06-30") %>% 
  filter(V_STERBEDATUM >"2019-06-30"| is.na(V_STERBEDATUM))

data1c<- data1c %>% 
  group_by(V_ID) %>% 
  arrange(START_DATE) %>% 
  mutate(n=row_number())

# nur erstes Intervall pro ID behalten 
data1d <- data1c %>% filter(n==1)

data1d <- data1d %>% dplyr::select(V_ID,V_PFLEGEGRAD ) %>% 
  rename(PFLEGEGRAD_END = V_PFLEGEGRAD)

# mit vorherigem Datensatz verbinden 
data1e <- left_join(data1b, data1d, by=c("V_ID"))
data1f <- data1e %>% filter(!(is.na(PFLEGEGRAD_END))) %>% 
  mutate(V_PFLEGEGRAD2 = as.numeric(V_PFLEGEGRAD),
         PFLEGEGRAD_END2 = as.numeric(PFLEGEGRAD_END),
         diff_PG = PFLEGEGRAD_END2-V_PFLEGEGRAD2,
         diff_PG =as.factor(diff_PG))

table1_pg_mitte <- table1(~ V_PFLEGEGRAD + PFLEGEGRAD_END + diff_PG | TG, data=data1f)


### Pflegegrad am 31.12.2019
data1c <- data1 %>% 
  filter(END_DATE>="2019-12-31") %>% 
  filter(is.na(V_STERBEDATUM))

data1d<- data1c 

data1d <- data1d %>% 
  dplyr::select(V_ID,V_PFLEGEGRAD ) %>% 
  rename(PFLEGEGRAD_END = V_PFLEGEGRAD)

# mit vorherigem Datensatz verbinden 
data1e <- left_join(data1b, data1d, by=c("V_ID"))
data1f <- data1e %>% filter(!(is.na(PFLEGEGRAD_END))) %>% 
  mutate(V_PFLEGEGRAD2 = as.numeric(V_PFLEGEGRAD),
         PFLEGEGRAD_END2 =as.numeric(PFLEGEGRAD_END),
         diff_PG = PFLEGEGRAD_END2-V_PFLEGEGRAD2,
         diff_PG = as.factor(diff_PG))

table1_pg_ende <- table1(~ V_PFLEGEGRAD + PFLEGEGRAD_END + diff_PG | TG, data=data1f)

list_of_datasets <- list("table1" = table1,
                         "table1_feb" = table1_feb,
                         "table1_pg_mitte" = table1_pg_mitte,
                         "table1_pg_ende" = table1_pg_ende)

write.xlsx(list_of_datasets , file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_Table1_2019.xlsx")
```

```{r descr_tg_2019, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(TG) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(TG) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("TG"))

tab1_tg <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```

```{r descr_pg_2019, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(V_PFLEGEGRAD) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(V_PFLEGEGRAD) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("V_PFLEGEGRAD"))

tab1_pg <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```


```{r descr_alter_2019, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(V_ALTER) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(V_ALTER) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("V_ALTER"))

tab1_age <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```


```{r descr_tod_2019, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(tod) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(tod) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("tod"))

tab1_tod <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```

```{r export_2019}
list_of_datasets <- list("tab1_alle" = tab1_alle,
                         "tab1_tg" = tab1_tg,
                         "tab1_pg" = tab1_pg,
                         "tab1_age" = tab1_age,
                         "tab1_tod" = tab1_tod)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_Tab1_2019.xlsx")
```


## weitere Tabellen des Abschnitts
```{r tab1_VZ_Tage_2019}
# ZG3
tab1_zg3_VZ_Tage <- data1 %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# ZG4
tab1_zg4_VZ_Tage <- data1 %>% 
  filter(ZG4 == 1 ) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG4", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# Daten verbinden
tab1_VZ_Tage <- bind_rows(tab1_zg3_VZ_Tage, tab1_zg4_VZ_Tage) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tab1_KH_Tage_2019}
# ZG3
tab1_zg3_KH_Tage <- data3a %>% 
  filter(ZG3==1 & !(is.na(kh_dauer))) %>%  
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# ZG4
tab1_zg4_KH_Tage <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG4", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# Daten verbinden
tab1_KH_Tage <- bind_rows(tab1_zg3_KH_Tage, tab1_zg4_KH_Tage) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```


```{r tab1_oKH_Tage_2019}
tab1_oKH_Tage <- left_join(tab1_VZ_Tage, tab1_KH_Tage, by=c("V_PFLEGEGRAD","ZG")) %>% 
  mutate(n_oKH_Tage = n_Tage,
         oKH_Tage_voll =  VZ_Tage_voll-KH_Tage_voll,
         oKH_Tage = oKH_Tage_voll/n_oKH_Tage) 
```

```{r tab1_n_AG_2019}
n_AG_zg3 <- data1 %>% filter(ZG3 == 1) %>%
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3")

n_AG_zg4<- data1 %>% filter(ZG4 == 1) %>%  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE)%>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG4")

tab1_n_AG <- bind_rows(n_AG_zg3,n_AG_zg4) %>% relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tabelle5_2019}
# Data prep
data1 <- data1 %>% mutate(tod = if_else(is.na(V_STERBEDATUM),0,1))
data3a <- data3a %>% mutate(tod = if_else(is.na(V_STERBEDATUM),0,1))

# ZG3 
tab5_zg3_bas <- data1 %>% 
  filter(ZG3==1 ) %>%
  group_by(TG,V_PFLEGEGRAD, V_ALTER,tod,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer ))

# ZG3-KH
tab5_zg3_kh <- data3a %>%
  filter(ZG3==1 & !(is.na(kh_dauer)))%>%
  group_by(TG,V_PFLEGEGRAD, V_ALTER,tod,.drop = FALSE)%>% 
  summarise(KH_tage= sum(kh_dauer))  

# Daten verbinden
tab5_zg3 <- left_join(tab5_zg3_bas, tab5_zg3_kh, by = c("TG", "V_PFLEGEGRAD", "V_ALTER", "tod"))

tab5 <- tab5_zg3 %>% 
  filter(!(is.na(tod))) %>% 
  mutate( KH_tage = if_else(is.na(KH_tage), 0, KH_tage),
          oKH_Tage = VZ_tage - KH_tage,
          per_oKH_Tage = KH_tage / VZ_tage)

VZ_ZG_tab5 <- tab5
```

```{r tab5_VZ_Tage_2019}
# ZG3
tab5_zg3_VZ_Tage <- data1 %>% filter(ZG3==1 ) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG3", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# ZG4
tab5_zg4_VZ_Tage <- data1 %>% 
  filter(ZG3==1 ) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG4", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# Daten verbinden
tab5_VZ_Tage <- bind_rows(tab5_zg3_VZ_Tage, tab5_zg4_VZ_Tage) %>% 
  relocate(ZG, .before = TG)
```


```{r tab5_KH_Tage_2019}
# ZG3
tab5_zg3_KH_Tage <- data3a %>% 
  filter(ZG3==1 & !(is.na(kh_dauer))) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG3", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# ZG4
tab5_zg4_KH_Tage <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>%
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG4", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# Daten verbinden
tab5_KH_Tage <- bind_rows(tab5_zg3_KH_Tage, tab5_zg4_KH_Tage) %>% relocate(ZG, .before = TG)
```

```{r tab5_oKH_Tage_2019}
tab5_oKH_Tage <- left_join(tab5_VZ_Tage,tab5_KH_Tage, by=c("TG", "V_PFLEGEGRAD","ZG")) %>% 
   mutate(n_oKH_Tage = n_Tage,
          oKH_Tage_voll =  VZ_Tage_voll-KH_Tage_voll,
          oKH_Tage = oKH_Tage_voll/n_oKH_Tage) 
```

```{r tab5_n_AG_2019}
n_AG_zg3 <- data1 %>% filter(ZG3==1 ) %>%  
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG3")

n_AG_zg4<- data1 %>% filter(ZG4==1 ) %>%  
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG4")

tab5_n_AG <- bind_rows(n_AG_zg3,n_AG_zg4) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tabelle9_2019}
# Data prep
# ZG3 
tab9_zg3_bas <- data1 %>%
  filter(ZG3 == 1) %>%
  group_by(TG,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  mutate(ZG = "ZG3") 

# ZG4
tab9_zg4_bas <- data1 %>% 
  filter(ZG4 == 1) %>%  
  group_by(TG,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  mutate(ZG = "ZG4") 

# ZG3-KH
tab9_zg3_kh <- data3a %>% 
  filter(ZG3 == 1 & !(is.na(kh_dauer)))%>%
  group_by(TG,.drop = FALSE)%>% 
  summarise(KH_tage = sum(kh_dauer))  %>% 
  mutate(ZG = "ZG3") 

# ZG4-KH
tab9_zg4_kh <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>% 
  group_by(TG,.drop = FALSE)%>% 
  summarise(KH_tage = sum(kh_dauer))  %>% 
  mutate(ZG = "ZG4") 

# Daten verbinden
tab9_zg3 <- left_join(tab9_zg3_bas, tab9_zg3_kh, by=c("TG","ZG"))
tab9_zg4 <- left_join(tab9_zg4_bas, tab9_zg4_kh, by=c("TG","ZG"))

tab9 <- bind_rows(tab9_zg3,tab9_zg4) %>% 
  mutate(oKH_Tage= VZ_tage-KH_tage,
         per_oKH_Tage= KH_tage/VZ_tage)

VZ_ZG_tab9 <- tab9 %>% relocate(ZG, .before = TG) 
```


```{r export2_2019}
# list_of_datasets <- list("VZ_ZG" = VZ_ZG,"tab1_VZ_Tage" =tab1_VZ_Tage,
#                          "tab1_KH_Tage" =tab1_KH_Tage,"tab1_oKH_Tage" =tab1_oKH_Tage,
#                          "tab1_n_AG" =tab1_n_AG)
# write.xlsx(list_of_datasets, 
#            file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/01_PG_2019.xlsx")


list_of_datasets <- list("VZ_ZG_tab5" = VZ_ZG_tab5,
                         "tab5_VZ_Tage" =tab5_VZ_Tage,
                         "tab5_KH_Tage" =tab5_KH_Tage,
                         "tab5_oKH_Tage" =tab5_oKH_Tage,
                         "tab5_n_AG" =tab5_n_AG)
write.xlsx(VZ_ZG_tab5, 
           file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/05_TG_PG_2019.xlsx")


list_of_datasets <- list("VZ_ZG_tab9" = VZ_ZG_tab9)
write.xlsx(list_of_datasets, 
           file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/01a_TG_2019.xlsx")
```










# 2020

```{r data_load_2020}
setwd("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/02_rdata/2020")

load("data11_neu.Rda")
load("kh11.Rda")

data1 <- data11_neu
```


```{r kh_prep_2020, include = FALSE}
### Data prep
# Dauer des Intervalls
data1<- data1 %>% 
  mutate(intervalldauer = as.double(END_DATE - START_DATE +1),
         tod = if_else(is.na(V_STERBEDATUM),0,1))

# KH-Daten vorbereiten
# nur vollstationär behalten
kh11<- kh11  %>%
  mutate(KH_AUF_ANLASS2 = str_pad(KH_AUF_ANLASS, 4, pad = "0"),
         KH_AUF_ANLASS_Beg = substr(KH_AUF_ANLASS2, 0, 2),
         KH_AUF_ANLASS_End = substr(KH_AUF_ANLASS2, 3, 4)) %>% 
  filter(KH_AUF_ANLASS_Beg=="01")

# überschneidende Hospitalisierungen verbinden
kh11$KH_BEGINN_DAT <- ymd(kh11$KH_BEGINN_DAT)
kh11$KH_ENDE_DAT <- ymd(kh11$KH_ENDE_DAT)

kh11a <- kh11 %>%
  arrange(V_ID, KH_BEGINN_DAT) %>% 
  group_by(V_ID) %>%
  mutate(indx = c(0, cumsum(as.numeric(lead(KH_BEGINN_DAT )) > cummax(as.numeric(KH_ENDE_DAT )))[-n()])) %>%
  group_by(V_ID, indx) %>%
  summarise(KH_BEGINN_DAT2 = min(KH_BEGINN_DAT), 
            KH_ENDE_DAT2 = max(KH_ENDE_DAT))

# Kosten und Uhrzeit des ersten Falles betrachten 
kh11b <- left_join(kh11a, kh11, by=c("V_ID", "KH_BEGINN_DAT2"="KH_BEGINN_DAT"))

# bei mehreren Fällen mit Beginn am Aufnahmetag, Fall mit längster Aufenthaltsdauer behalten
kh11c <- kh11b %>% 
  group_by(V_ID, KH_BEGINN_DAT2) %>% 
  arrange(desc(KH_ENDE_DAT)) %>% 
  mutate(n = paste0( row_number())) %>% 
  filter(n==1)

# Stammdaten mit Krankenhausdaten verbinden
data3 <- left_join(data1, kh11c, by=("V_ID"))

data3a <- data3 %>% 
  filter(!(KH_BEGINN_DAT2 >= END_DATE | KH_ENDE_DAT2 <= START_DATE)) %>% 
  rowwise() %>% 
  mutate(rel_KH_BEGINN_DAT = max(KH_BEGINN_DAT2,START_DATE),
         rel_KH_ENDE_DAT = min(KH_ENDE_DAT2,END_DATE)) %>% 
  dplyr::mutate(kh_dauer = as.double(rel_KH_ENDE_DAT - rel_KH_BEGINN_DAT + 1))

# Faktorvariablen für group_by (drop false) notwendig
data1 <- data1 %>% mutate(tg = as.factor(tg), 
                          V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD), 
                          V_ALTER = as.factor(V_ALTER)) 

data3a <- data3a %>% 
  mutate(tg = as.factor(tg), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD), 
         V_ALTER = as.factor(V_ALTER)) 
```



```{r verweildauer_2020}
verweildauer_zg3 <- data3a %>% 
  filter(ZG3==1) %>% 
  group_by(V_ID) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4 <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_TG  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,TG) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_TG <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,TG) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_pg  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_pg <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_pg_TG  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_pg_TG <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_pg_TG_tod <- data3a %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD, V_ALTER,tod) %>% 
  summarise(Verweildauer = sum(kh_dauer))

list_of_datasets <- list("verweildauer_zg3" = verweildauer_zg3,
                         "verweildauer_zg4" = verweildauer_zg4,
                         "verweildauer_zg3_TG" = verweildauer_zg3_TG, 
                         "verweildauer_zg4_TG" = verweildauer_zg4_TG,
                         "verweildauer_zg3_pg" = verweildauer_zg3_pg, 
                         "verweildauer_zg4_pg" = verweildauer_zg4_pg,
                         "verweildauer_zg3_pg_TG" = verweildauer_zg3_pg_TG, 
                         "verweildauer_zg4_pg_TG" = verweildauer_zg4_pg_TG,
                         "verweildauer_pg_TG_tod" =verweildauer_pg_TG_tod)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2020.xlsx")
```
```{r descr_alle_2020, include = FALSE}
# Dauer des Intervalls
data1<- data1 %>% 
  mutate(intervalldauer = as.double(END_DATE - START_DATE +1),
         tod = if_else(is.na(V_STERBEDATUM),0,1))

# ZG3 
tab1_zg3_bas <- data1 %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer)) 

# ZG3-KH
tab1_zg3_kh <- data3a %>%
  ungroup() %>% 
  summarise(KH_tage= sum(kh_dauer)) 

# Daten verbinden
tab1_zg3 <- bind_cols(tab1_zg3_bas, tab1_zg3_kh)

tab1_alle <- tab1_zg3 %>% 
  mutate(oKH_tage= VZ_tage-KH_tage,
         VZ_tage_N = VZ_tage/N,
         KH_tage_N = KH_tage/N,
         oKH_tage_N = oKH_tage/N,
         per_KH_Tage= KH_tage/VZ_tage)

data1a <- data1 %>% 
  group_by(V_ID) %>% 
  arrange(START_DATE) %>% 
  mutate(n=row_number())

# nur erstes Intervall pro ID behalten
table(data1a$n)
data1b <- data1a %>% filter(n==1)

data1b$V_ALTER<- 
  factor(data1b$V_ALTER , 
         levels = c(3,4,5,6,7,8,9),
         labels = c("60-64","65-69","70-74","75-79","80-84","85-89","90 und mehr" ))

label(data1b$V_ALTER) <- "Altersgruppe"

data1b$tod<- 
  factor(data1b$tod , 
         levels=c(0,1),
         labels=c("nein", "ja"))

label(data1b$tod) <- "Tod"

label(data1b$V_PFLEGEGRAD) <- "Pflegegrad"

table1(~ V_ALTER + tod + V_PFLEGEGRAD| TG, data=data1b)
table1 <- table1(~ V_ALTER + tod + V_PFLEGEGRAD| TG, data=data1b)
table1 <- as.data.frame(table1)




### Pflegegrad am 30.06.2020 (ehemals 31.12.2020)
data1c <- data1 %>% 
  filter(END_DATE >= "2020-06-30") %>% 
  filter(V_STERBEDATUM >"2020-06-30"| is.na(V_STERBEDATUM))

data1c<- data1c %>% 
  group_by(V_ID) %>% 
  arrange(START_DATE) %>% 
  mutate(n=row_number())

# nur erstes Intervall pro ID behalten 
data1d <- data1c %>% filter(n==1)

data1d <- data1d %>% dplyr::select(V_ID,V_PFLEGEGRAD ) %>% 
  rename(PFLEGEGRAD_END = V_PFLEGEGRAD)

# mit vorherigem Datensatz verbinden 
data1e <- left_join(data1b, data1d, by=c("V_ID"))
data1f <- data1e %>% filter(!(is.na(PFLEGEGRAD_END))) %>% 
  mutate(V_PFLEGEGRAD2=as.numeric(V_PFLEGEGRAD),
         PFLEGEGRAD_END2=as.numeric(PFLEGEGRAD_END),
         diff_PG= PFLEGEGRAD_END2-V_PFLEGEGRAD2,
         diff_PG=as.factor(diff_PG))

table1_pg_mitte <- table1(~ V_PFLEGEGRAD + PFLEGEGRAD_END + diff_PG | TG, data=data1f)


### Pflegegrad am 31.12.2020
data1c <- data1 %>% 
  filter(END_DATE>="2020-12-31") %>% 
  filter(is.na(V_STERBEDATUM))

data1d<- data1c 

data1d <- data1d %>% 
  dplyr::select(V_ID,V_PFLEGEGRAD ) %>% 
  rename(PFLEGEGRAD_END = V_PFLEGEGRAD)

# mit vorherigem Datensatz verbinden 
data1e <- left_join(data1b, data1d, by=c("V_ID"))
data1f <- data1e %>% filter(!(is.na(PFLEGEGRAD_END))) %>% 
  mutate(V_PFLEGEGRAD2 = as.numeric(V_PFLEGEGRAD),
         PFLEGEGRAD_END2 =as.numeric(PFLEGEGRAD_END),
         diff_PG = PFLEGEGRAD_END2-V_PFLEGEGRAD2,
         diff_PG = as.factor(diff_PG))

table1_pg_ende <- table1(~ V_PFLEGEGRAD + PFLEGEGRAD_END + diff_PG | TG, data=data1f)

list_of_datasets <- list("table1" = table1,
                         "table1_pg_mitte" = table1_pg_mitte,
                         "table1_pg_ende" = table1_pg_ende)

write.xlsx(list_of_datasets , file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_Table1_2020.xlsx")
```

```{r descr_tg_2020, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(TG) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(TG) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("TG"))

tab1_tg <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```

```{r descr_pg_2020, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(V_PFLEGEGRAD) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(V_PFLEGEGRAD) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("V_PFLEGEGRAD"))

tab1_pg <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```


```{r descr_alter_2020, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(V_ALTER) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(V_ALTER) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("V_ALTER"))

tab1_age <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```


```{r descr_tod_2020, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(tod) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(tod) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("tod"))

tab1_tod <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```

```{r export_2020}
list_of_datasets <- list("tab1_alle" = tab1_alle,
                         "tab1_tg" = tab1_tg,
                         "tab1_pg" = tab1_pg,
                         "tab1_age" = tab1_age,
                         "tab1_tod" = tab1_tod)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_Tab1_2020.xlsx")
```


## weitere Tabellen des Abschnitts
```{r tab1_VZ_Tage_2020}
# ZG3
tab1_zg3_VZ_Tage <- data1 %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# ZG4
tab1_zg4_VZ_Tage <- data1 %>% 
  filter(ZG4 == 1 ) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG4", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# Daten verbinden
tab1_VZ_Tage <- bind_rows(tab1_zg3_VZ_Tage, tab1_zg4_VZ_Tage) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tab1_KH_Tage_2020}
# ZG3
tab1_zg3_KH_Tage <- data3a %>% 
  filter(ZG3==1 & !(is.na(kh_dauer))) %>%  
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# ZG4
tab1_zg4_KH_Tage <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG4", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# Daten verbinden
tab1_KH_Tage <- bind_rows(tab1_zg3_KH_Tage, tab1_zg4_KH_Tage) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```


```{r tab1_oKH_Tage_2020}
tab1_oKH_Tage <- left_join(tab1_VZ_Tage, tab1_KH_Tage, by=c("V_PFLEGEGRAD","ZG")) %>% 
  mutate(n_oKH_Tage = n_Tage,
         oKH_Tage_voll =  VZ_Tage_voll-KH_Tage_voll,
         oKH_Tage = oKH_Tage_voll/n_oKH_Tage) 
```

```{r tab1_n_AG_2020}
n_AG_zg3 <- data1 %>% filter(ZG3 == 1) %>%
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3")

n_AG_zg4<- data1 %>% filter(ZG4 == 1) %>%  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE)%>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG4")

tab1_n_AG <- bind_rows(n_AG_zg3,n_AG_zg4) %>% relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tabelle5_2020}
# Data prep
data1 <- data1 %>% mutate(tod = if_else(is.na(V_STERBEDATUM),0,1))
data3a <- data3a %>% mutate(tod = if_else(is.na(V_STERBEDATUM),0,1))

# ZG3 
tab5_zg3_bas <- data1 %>% 
  filter(ZG3==1 ) %>%
  group_by(TG,V_PFLEGEGRAD, V_ALTER,tod,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer ))

# ZG3-KH
tab5_zg3_kh <- data3a %>%
  filter(ZG3==1 & !(is.na(kh_dauer)))%>%
  group_by(TG,V_PFLEGEGRAD, V_ALTER,tod,.drop = FALSE)%>% 
  summarise(KH_tage= sum(kh_dauer))  

# Daten verbinden
tab5_zg3 <- left_join(tab5_zg3_bas, tab5_zg3_kh, by = c("TG", "V_PFLEGEGRAD", "V_ALTER", "tod"))

tab5 <- tab5_zg3 %>% 
  filter(!(is.na(tod))) %>% 
  mutate( KH_tage = if_else(is.na(KH_tage), 0, KH_tage),
          oKH_Tage = VZ_tage - KH_tage,
          per_oKH_Tage = KH_tage / VZ_tage)

VZ_ZG_tab5 <- tab5
```

```{r tab5_VZ_Tage_2020}
# ZG3
tab5_zg3_VZ_Tage <- data1 %>% filter(ZG3==1 ) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG3", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# ZG4
tab5_zg4_VZ_Tage <- data1 %>% 
  filter(ZG3==1 ) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG4", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# Daten verbinden
tab5_VZ_Tage <- bind_rows(tab5_zg3_VZ_Tage, tab5_zg4_VZ_Tage) %>% 
  relocate(ZG, .before = TG)
```


```{r tab5_KH_Tage_2020}
# ZG3
tab5_zg3_KH_Tage <- data3a %>% 
  filter(ZG3==1 & !(is.na(kh_dauer))) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG3", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# ZG4
tab5_zg4_KH_Tage <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>%
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG4", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# Daten verbinden
tab5_KH_Tage <- bind_rows(tab5_zg3_KH_Tage, tab5_zg4_KH_Tage) %>% relocate(ZG, .before = TG)
```

```{r tab5_oKH_Tage_2020}
tab5_oKH_Tage <- left_join(tab5_VZ_Tage,tab5_KH_Tage, by=c("TG", "V_PFLEGEGRAD","ZG")) %>% 
   mutate(n_oKH_Tage = n_Tage,
          oKH_Tage_voll =  VZ_Tage_voll-KH_Tage_voll,
          oKH_Tage = oKH_Tage_voll/n_oKH_Tage) 
```

```{r tab5_n_AG_2020}
n_AG_zg3 <- data1 %>% filter(ZG3==1 ) %>%  
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG3")

n_AG_zg4<- data1 %>% filter(ZG4==1 ) %>%  
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG4")

tab5_n_AG <- bind_rows(n_AG_zg3,n_AG_zg4) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tabelle9_2020}
# Data prep
# ZG3 
tab9_zg3_bas <- data1 %>%
  filter(ZG3 == 1) %>%
  group_by(TG,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  mutate(ZG = "ZG3") 

# ZG4
tab9_zg4_bas <- data1 %>% 
  filter(ZG4 == 1) %>%  
  group_by(TG,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  mutate(ZG = "ZG4") 

# ZG3-KH
tab9_zg3_kh <- data3a %>% 
  filter(ZG3 == 1 & !(is.na(kh_dauer)))%>%
  group_by(TG,.drop = FALSE)%>% 
  summarise(KH_tage = sum(kh_dauer))  %>% 
  mutate(ZG = "ZG3") 

# ZG4-KH
tab9_zg4_kh <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>% 
  group_by(TG,.drop = FALSE)%>% 
  summarise(KH_tage = sum(kh_dauer))  %>% 
  mutate(ZG = "ZG4") 

# Daten verbinden
tab9_zg3 <- left_join(tab9_zg3_bas, tab9_zg3_kh, by=c("TG","ZG"))
tab9_zg4 <- left_join(tab9_zg4_bas, tab9_zg4_kh, by=c("TG","ZG"))

tab9 <- bind_rows(tab9_zg3,tab9_zg4) %>% 
  mutate(oKH_Tage= VZ_tage-KH_tage,
         per_oKH_Tage= KH_tage/VZ_tage)

VZ_ZG_tab9 <- tab9 %>% relocate(ZG, .before = TG) 
```


```{r export2_2020}
# list_of_datasets <- list("VZ_ZG" = VZ_ZG,"tab1_VZ_Tage" =tab1_VZ_Tage,
#                          "tab1_KH_Tage" =tab1_KH_Tage,"tab1_oKH_Tage" =tab1_oKH_Tage,
#                          "tab1_n_AG" =tab1_n_AG)
# write.xlsx(list_of_datasets, 
#            file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/01_PG_2020.xlsx")


list_of_datasets <- list("VZ_ZG_tab5" = VZ_ZG_tab5,
                         "tab5_VZ_Tage" =tab5_VZ_Tage,
                         "tab5_KH_Tage" =tab5_KH_Tage,
                         "tab5_oKH_Tage" =tab5_oKH_Tage,
                         "tab5_n_AG" =tab5_n_AG)
write.xlsx(VZ_ZG_tab5, 
           file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/05_TG_PG_2020.xlsx")


list_of_datasets <- list("VZ_ZG_tab9" = VZ_ZG_tab9)
write.xlsx(list_of_datasets, 
           file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/01a_TG_2020.xlsx")
```










# 2021

```{r data_load_2021}
setwd("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/02_rdata/2021")

load("data11_neu.Rda")
load("kh11.Rda")

data1 <- data11_neu
```


```{r kh_prep_2021, include = FALSE}
### Data prep
# Dauer des Intervalls
data1<- data1 %>% 
  mutate(intervalldauer = as.double(END_DATE - START_DATE +1),
         tod = if_else(is.na(V_STERBEDATUM),0,1))

# KH-Daten vorbereiten
# nur vollstationär behalten
kh11<- kh11  %>%
  mutate(KH_AUF_ANLASS2 = str_pad(KH_AUF_ANLASS, 4, pad = "0"),
         KH_AUF_ANLASS_Beg = substr(KH_AUF_ANLASS2, 0, 2),
         KH_AUF_ANLASS_End = substr(KH_AUF_ANLASS2, 3, 4)) %>% 
  filter(KH_AUF_ANLASS_Beg=="01")

# überschneidende Hospitalisierungen verbinden
kh11$KH_BEGINN_DAT <- ymd(kh11$KH_BEGINN_DAT)
kh11$KH_ENDE_DAT <- ymd(kh11$KH_ENDE_DAT)

kh11a <- kh11 %>%
  arrange(V_ID, KH_BEGINN_DAT) %>% 
  group_by(V_ID) %>%
  mutate(indx = c(0, cumsum(as.numeric(lead(KH_BEGINN_DAT )) > cummax(as.numeric(KH_ENDE_DAT )))[-n()])) %>%
  group_by(V_ID, indx) %>%
  summarise(KH_BEGINN_DAT2 = min(KH_BEGINN_DAT), 
            KH_ENDE_DAT2 = max(KH_ENDE_DAT))

# Kosten und Uhrzeit des ersten Falles betrachten 
kh11b <- left_join(kh11a, kh11, by=c("V_ID", "KH_BEGINN_DAT2"="KH_BEGINN_DAT"))

# bei mehreren Fällen mit Beginn am Aufnahmetag, Fall mit längster Aufenthaltsdauer behalten
kh11c <- kh11b %>% 
  group_by(V_ID, KH_BEGINN_DAT2) %>% 
  arrange(desc(KH_ENDE_DAT)) %>% 
  mutate(n = paste0( row_number())) %>% 
  filter(n==1)

# Stammdaten mit Krankenhausdaten verbinden
data3 <- left_join(data1, kh11c, by=("V_ID"))

data3a <- data3 %>% 
  filter(!(KH_BEGINN_DAT2 >= END_DATE | KH_ENDE_DAT2 <= START_DATE)) %>% 
  rowwise() %>% 
  mutate(rel_KH_BEGINN_DAT = max(KH_BEGINN_DAT2,START_DATE),
         rel_KH_ENDE_DAT = min(KH_ENDE_DAT2,END_DATE)) %>% 
  dplyr::mutate(kh_dauer = as.double(rel_KH_ENDE_DAT - rel_KH_BEGINN_DAT + 1))

# Faktorvariablen für group_by (drop false) notwendig
data1 <- data1 %>% mutate(tg = as.factor(tg), 
                          V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD), 
                          V_ALTER = as.factor(V_ALTER)) 

data3a <- data3a %>% 
  mutate(tg = as.factor(tg), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD), 
         V_ALTER = as.factor(V_ALTER)) 
```



```{r verweildauer_2021}
verweildauer_zg3 <- data3a %>% 
  filter(ZG3==1) %>% 
  group_by(V_ID) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4 <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_TG  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,TG) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_TG <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,TG) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_pg  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_pg <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_pg_TG  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_pg_TG <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_pg_TG_tod <- data3a %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD, V_ALTER,tod) %>% 
  summarise(Verweildauer = sum(kh_dauer))

list_of_datasets <- list("verweildauer_zg3" = verweildauer_zg3,
                         "verweildauer_zg4" = verweildauer_zg4,
                         "verweildauer_zg3_TG" = verweildauer_zg3_TG, 
                         "verweildauer_zg4_TG" = verweildauer_zg4_TG,
                         "verweildauer_zg3_pg" = verweildauer_zg3_pg, 
                         "verweildauer_zg4_pg" = verweildauer_zg4_pg,
                         "verweildauer_zg3_pg_TG" = verweildauer_zg3_pg_TG, 
                         "verweildauer_zg4_pg_TG" = verweildauer_zg4_pg_TG,
                         "verweildauer_pg_TG_tod" =verweildauer_pg_TG_tod)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2021.xlsx")
```
```{r descr_alle_2021, include = FALSE}
# Dauer des Intervalls
data1<- data1 %>% 
  mutate(intervalldauer = as.double(END_DATE - START_DATE +1),
         tod = if_else(is.na(V_STERBEDATUM),0,1))

# ZG3 
tab1_zg3_bas <- data1 %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer)) 

# ZG3-KH
tab1_zg3_kh <- data3a %>%
  ungroup() %>% 
  summarise(KH_tage= sum(kh_dauer)) 

# Daten verbinden
tab1_zg3 <- bind_cols(tab1_zg3_bas, tab1_zg3_kh)

tab1_alle <- tab1_zg3 %>% 
  mutate(oKH_tage= VZ_tage-KH_tage,
         VZ_tage_N = VZ_tage/N,
         KH_tage_N = KH_tage/N,
         oKH_tage_N = oKH_tage/N,
         per_KH_Tage= KH_tage/VZ_tage)

data1a <- data1 %>% 
  group_by(V_ID) %>% 
  arrange(START_DATE) %>% 
  mutate(n=row_number())

# nur erstes Intervall pro ID behalten
table(data1a$n)
data1b <- data1a %>% filter(n==1)

data1b$V_ALTER<- 
  factor(data1b$V_ALTER , 
         levels = c(3,4,5,6,7,8,9),
         labels = c("60-64","65-69","70-74","75-79","80-84","85-89","90 und mehr" ))

label(data1b$V_ALTER) <- "Altersgruppe"

data1b$tod<- 
  factor(data1b$tod , 
         levels=c(0,1),
         labels=c("nein", "ja"))

label(data1b$tod) <- "Tod"

label(data1b$V_PFLEGEGRAD) <- "Pflegegrad"

table1(~ V_ALTER + tod + V_PFLEGEGRAD| TG, data=data1b)
table1 <- table1(~ V_ALTER + tod + V_PFLEGEGRAD| TG, data=data1b)
table1 <- as.data.frame(table1)




### Pflegegrad am 30.06.2021 (ehemals 31.12.2021)
data1c <- data1 %>% 
  filter(END_DATE >= "2021-06-30") %>% 
  filter(V_STERBEDATUM >"2021-06-30"| is.na(V_STERBEDATUM))

data1c<- data1c %>% 
  group_by(V_ID) %>% 
  arrange(START_DATE) %>% 
  mutate(n=row_number())

# nur erstes Intervall pro ID behalten 
data1d <- data1c %>% filter(n==1)

data1d <- data1d %>% dplyr::select(V_ID,V_PFLEGEGRAD ) %>% 
  rename(PFLEGEGRAD_END = V_PFLEGEGRAD)

# mit vorherigem Datensatz verbinden 
data1e <- left_join(data1b, data1d, by=c("V_ID"))
data1f <- data1e %>% filter(!(is.na(PFLEGEGRAD_END))) %>% 
  mutate(V_PFLEGEGRAD2=as.numeric(V_PFLEGEGRAD),
         PFLEGEGRAD_END2=as.numeric(PFLEGEGRAD_END),
         diff_PG= PFLEGEGRAD_END2-V_PFLEGEGRAD2,
         diff_PG=as.factor(diff_PG))

table1_pg_mitte <- table1(~ V_PFLEGEGRAD + PFLEGEGRAD_END + diff_PG | TG, data=data1f)


### Pflegegrad am 31.12.2021
data1c <- data1 %>% 
  filter(END_DATE>="2021-12-31") %>% 
  filter(is.na(V_STERBEDATUM))

data1d<- data1c 

data1d <- data1d %>% 
  dplyr::select(V_ID,V_PFLEGEGRAD ) %>% 
  rename(PFLEGEGRAD_END = V_PFLEGEGRAD)

# mit vorherigem Datensatz verbinden 
data1e <- left_join(data1b, data1d, by=c("V_ID"))
data1f <- data1e %>% filter(!(is.na(PFLEGEGRAD_END))) %>% 
  mutate(V_PFLEGEGRAD2 = as.numeric(V_PFLEGEGRAD),
         PFLEGEGRAD_END2 =as.numeric(PFLEGEGRAD_END),
         diff_PG = PFLEGEGRAD_END2-V_PFLEGEGRAD2,
         diff_PG = as.factor(diff_PG))

table1_pg_ende <- table1(~ V_PFLEGEGRAD + PFLEGEGRAD_END + diff_PG | TG, data=data1f)

list_of_datasets <- list("table1" = table1,
                         "table1_pg_mitte" = table1_pg_mitte,
                         "table1_pg_ende" = table1_pg_ende)

write.xlsx(list_of_datasets , file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_Table1_2021.xlsx")
```

```{r descr_tg_2021, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(TG) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(TG) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("TG"))

tab1_tg <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```

```{r descr_pg_2021, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(V_PFLEGEGRAD) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(V_PFLEGEGRAD) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("V_PFLEGEGRAD"))

tab1_pg <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```


```{r descr_alter_2021, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(V_ALTER) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(V_ALTER) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("V_ALTER"))

tab1_age <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```


```{r descr_tod_2021, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(tod) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(tod) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("tod"))

tab1_tod <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```

```{r export_2021}
list_of_datasets <- list("tab1_alle" = tab1_alle,
                         "tab1_tg" = tab1_tg,
                         "tab1_pg" = tab1_pg,
                         "tab1_age" = tab1_age,
                         "tab1_tod" = tab1_tod)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_Tab1_2021.xlsx")
```


## weitere Tabellen des Abschnitts
```{r tab1_VZ_Tage_2021}
# ZG3
tab1_zg3_VZ_Tage <- data1 %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# ZG4
tab1_zg4_VZ_Tage <- data1 %>% 
  filter(ZG4 == 1 ) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG4", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# Daten verbinden
tab1_VZ_Tage <- bind_rows(tab1_zg3_VZ_Tage, tab1_zg4_VZ_Tage) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tab1_KH_Tage_2021}
# ZG3
tab1_zg3_KH_Tage <- data3a %>% 
  filter(ZG3==1 & !(is.na(kh_dauer))) %>%  
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# ZG4
tab1_zg4_KH_Tage <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG4", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# Daten verbinden
tab1_KH_Tage <- bind_rows(tab1_zg3_KH_Tage, tab1_zg4_KH_Tage) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```


```{r tab1_oKH_Tage_2021}
tab1_oKH_Tage <- left_join(tab1_VZ_Tage, tab1_KH_Tage, by=c("V_PFLEGEGRAD","ZG")) %>% 
  mutate(n_oKH_Tage = n_Tage,
         oKH_Tage_voll =  VZ_Tage_voll-KH_Tage_voll,
         oKH_Tage = oKH_Tage_voll/n_oKH_Tage) 
```

```{r tab1_n_AG_2021}
n_AG_zg3 <- data1 %>% filter(ZG3 == 1) %>%
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3")

n_AG_zg4<- data1 %>% filter(ZG4 == 1) %>%  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE)%>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG4")

tab1_n_AG <- bind_rows(n_AG_zg3,n_AG_zg4) %>% relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tabelle5_2021}
# Data prep
data1 <- data1 %>% mutate(tod = if_else(is.na(V_STERBEDATUM),0,1))
data3a <- data3a %>% mutate(tod = if_else(is.na(V_STERBEDATUM),0,1))

# ZG3 
tab5_zg3_bas <- data1 %>% 
  filter(ZG3==1 ) %>%
  group_by(TG,V_PFLEGEGRAD, V_ALTER,tod,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer ))

# ZG3-KH
tab5_zg3_kh <- data3a %>%
  filter(ZG3==1 & !(is.na(kh_dauer)))%>%
  group_by(TG,V_PFLEGEGRAD, V_ALTER,tod,.drop = FALSE)%>% 
  summarise(KH_tage= sum(kh_dauer))  

# Daten verbinden
tab5_zg3 <- left_join(tab5_zg3_bas, tab5_zg3_kh, by = c("TG", "V_PFLEGEGRAD", "V_ALTER", "tod"))

tab5 <- tab5_zg3 %>% 
  filter(!(is.na(tod))) %>% 
  mutate( KH_tage = if_else(is.na(KH_tage), 0, KH_tage),
          oKH_Tage = VZ_tage - KH_tage,
          per_oKH_Tage = KH_tage / VZ_tage)

VZ_ZG_tab5 <- tab5
```

```{r tab5_VZ_Tage_2021}
# ZG3
tab5_zg3_VZ_Tage <- data1 %>% filter(ZG3==1 ) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG3", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# ZG4
tab5_zg4_VZ_Tage <- data1 %>% 
  filter(ZG3==1 ) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG4", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# Daten verbinden
tab5_VZ_Tage <- bind_rows(tab5_zg3_VZ_Tage, tab5_zg4_VZ_Tage) %>% 
  relocate(ZG, .before = TG)
```


```{r tab5_KH_Tage_2021}
# ZG3
tab5_zg3_KH_Tage <- data3a %>% 
  filter(ZG3==1 & !(is.na(kh_dauer))) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG3", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# ZG4
tab5_zg4_KH_Tage <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>%
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG4", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# Daten verbinden
tab5_KH_Tage <- bind_rows(tab5_zg3_KH_Tage, tab5_zg4_KH_Tage) %>% relocate(ZG, .before = TG)
```

```{r tab5_oKH_Tage_2021}
tab5_oKH_Tage <- left_join(tab5_VZ_Tage,tab5_KH_Tage, by=c("TG", "V_PFLEGEGRAD","ZG")) %>% 
   mutate(n_oKH_Tage = n_Tage,
          oKH_Tage_voll =  VZ_Tage_voll-KH_Tage_voll,
          oKH_Tage = oKH_Tage_voll/n_oKH_Tage) 
```

```{r tab5_n_AG_2021}
n_AG_zg3 <- data1 %>% filter(ZG3==1 ) %>%  
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG3")

n_AG_zg4<- data1 %>% filter(ZG4==1 ) %>%  
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG4")

tab5_n_AG <- bind_rows(n_AG_zg3,n_AG_zg4) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tabelle9_2021}
# Data prep
# ZG3 
tab9_zg3_bas <- data1 %>%
  filter(ZG3 == 1) %>%
  group_by(TG,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  mutate(ZG = "ZG3") 

# ZG4
tab9_zg4_bas <- data1 %>% 
  filter(ZG4 == 1) %>%  
  group_by(TG,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  mutate(ZG = "ZG4") 

# ZG3-KH
tab9_zg3_kh <- data3a %>% 
  filter(ZG3 == 1 & !(is.na(kh_dauer)))%>%
  group_by(TG,.drop = FALSE)%>% 
  summarise(KH_tage = sum(kh_dauer))  %>% 
  mutate(ZG = "ZG3") 

# ZG4-KH
tab9_zg4_kh <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>% 
  group_by(TG,.drop = FALSE)%>% 
  summarise(KH_tage = sum(kh_dauer))  %>% 
  mutate(ZG = "ZG4") 

# Daten verbinden
tab9_zg3 <- left_join(tab9_zg3_bas, tab9_zg3_kh, by=c("TG","ZG"))
tab9_zg4 <- left_join(tab9_zg4_bas, tab9_zg4_kh, by=c("TG","ZG"))

tab9 <- bind_rows(tab9_zg3,tab9_zg4) %>% 
  mutate(oKH_Tage= VZ_tage-KH_tage,
         per_oKH_Tage= KH_tage/VZ_tage)

VZ_ZG_tab9 <- tab9 %>% relocate(ZG, .before = TG) 
```


```{r export2_2021}
# list_of_datasets <- list("VZ_ZG" = VZ_ZG,"tab1_VZ_Tage" =tab1_VZ_Tage,
#                          "tab1_KH_Tage" =tab1_KH_Tage,"tab1_oKH_Tage" =tab1_oKH_Tage,
#                          "tab1_n_AG" =tab1_n_AG)
# write.xlsx(list_of_datasets, 
#            file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/01_PG_2021.xlsx")


list_of_datasets <- list("VZ_ZG_tab5" = VZ_ZG_tab5,
                         "tab5_VZ_Tage" =tab5_VZ_Tage,
                         "tab5_KH_Tage" =tab5_KH_Tage,
                         "tab5_oKH_Tage" =tab5_oKH_Tage,
                         "tab5_n_AG" =tab5_n_AG)
write.xlsx(VZ_ZG_tab5, 
           file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/05_TG_PG_2021.xlsx")


list_of_datasets <- list("VZ_ZG_tab9" = VZ_ZG_tab9)
write.xlsx(list_of_datasets, 
           file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/01a_TG_2021.xlsx")
```




# 2022

```{r data_load_2022}
setwd("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/02_rdata/2022")

load("data11_neu.Rda")
load("kh11.Rda")

data1 <- data11_neu
```


```{r kh_prep_2022, include = FALSE}
### Data prep
# Dauer des Intervalls
data1<- data1 %>% 
  mutate(intervalldauer = as.double(END_DATE - START_DATE +1),
         tod = if_else(is.na(V_STERBEDATUM),0,1))

# KH-Daten vorbereiten
# nur vollstationär behalten
kh11<- kh11  %>%
  mutate(KH_AUF_ANLASS2 = str_pad(KH_AUF_ANLASS, 4, pad = "0"),
         KH_AUF_ANLASS_Beg = substr(KH_AUF_ANLASS2, 0, 2),
         KH_AUF_ANLASS_End = substr(KH_AUF_ANLASS2, 3, 4)) %>% 
  filter(KH_AUF_ANLASS_Beg=="01")

# überschneidende Hospitalisierungen verbinden
kh11$KH_BEGINN_DAT <- ymd(kh11$KH_BEGINN_DAT)
kh11$KH_ENDE_DAT <- ymd(kh11$KH_ENDE_DAT)

kh11a <- kh11 %>%
  arrange(V_ID, KH_BEGINN_DAT) %>% 
  group_by(V_ID) %>%
  mutate(indx = c(0, cumsum(as.numeric(lead(KH_BEGINN_DAT )) > cummax(as.numeric(KH_ENDE_DAT )))[-n()])) %>%
  group_by(V_ID, indx) %>%
  summarise(KH_BEGINN_DAT2 = min(KH_BEGINN_DAT), 
            KH_ENDE_DAT2 = max(KH_ENDE_DAT))

# Kosten und Uhrzeit des ersten Falles betrachten 
kh11b <- left_join(kh11a, kh11, by=c("V_ID", "KH_BEGINN_DAT2"="KH_BEGINN_DAT"))

# bei mehreren Fällen mit Beginn am Aufnahmetag, Fall mit längster Aufenthaltsdauer behalten
kh11c <- kh11b %>% 
  group_by(V_ID, KH_BEGINN_DAT2) %>% 
  arrange(desc(KH_ENDE_DAT)) %>% 
  mutate(n = paste0( row_number())) %>% 
  filter(n==1)

# Stammdaten mit Krankenhausdaten verbinden
data3 <- left_join(data1, kh11c, by=("V_ID"))

data3a <- data3 %>% 
  filter(!(KH_BEGINN_DAT2 >= END_DATE | KH_ENDE_DAT2 <= START_DATE)) %>% 
  rowwise() %>% 
  mutate(rel_KH_BEGINN_DAT = max(KH_BEGINN_DAT2,START_DATE),
         rel_KH_ENDE_DAT = min(KH_ENDE_DAT2,END_DATE)) %>% 
  dplyr::mutate(kh_dauer = as.double(rel_KH_ENDE_DAT - rel_KH_BEGINN_DAT + 1))

# Faktorvariablen für group_by (drop false) notwendig
data1 <- data1 %>% mutate(tg = as.factor(tg), 
                          V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD), 
                          V_ALTER = as.factor(V_ALTER)) 

data3a <- data3a %>% 
  mutate(tg = as.factor(tg), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD), 
         V_ALTER = as.factor(V_ALTER)) 
```



```{r verweildauer_2022}
verweildauer_zg3 <- data3a %>% 
  filter(ZG3==1) %>% 
  group_by(V_ID) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4 <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_TG  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,TG) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_TG <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,TG) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_pg  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_pg <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg3_pg_TG  <- data3a %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_zg4_pg_TG <- data3a %>% 
  filter(ZG4==1 ) %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD) %>% 
  summarise(Verweildauer = sum(kh_dauer))

verweildauer_pg_TG_tod <- data3a %>% 
  group_by(V_ID,TG,V_PFLEGEGRAD, V_ALTER,tod) %>% 
  summarise(Verweildauer = sum(kh_dauer))

list_of_datasets <- list("verweildauer_zg3" = verweildauer_zg3,
                         "verweildauer_zg4" = verweildauer_zg4,
                         "verweildauer_zg3_TG" = verweildauer_zg3_TG, 
                         "verweildauer_zg4_TG" = verweildauer_zg4_TG,
                         "verweildauer_zg3_pg" = verweildauer_zg3_pg, 
                         "verweildauer_zg4_pg" = verweildauer_zg4_pg,
                         "verweildauer_zg3_pg_TG" = verweildauer_zg3_pg_TG, 
                         "verweildauer_zg4_pg_TG" = verweildauer_zg4_pg_TG,
                         "verweildauer_pg_TG_tod" =verweildauer_pg_TG_tod)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2022.xlsx")
```
```{r descr_alle_2022, include = FALSE}
# Dauer des Intervalls
data1<- data1 %>% 
  mutate(intervalldauer = as.double(END_DATE - START_DATE +1),
         tod = if_else(is.na(V_STERBEDATUM),0,1))

# ZG3 
tab1_zg3_bas <- data1 %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer)) 

# ZG3-KH
tab1_zg3_kh <- data3a %>%
  ungroup() %>% 
  summarise(KH_tage= sum(kh_dauer)) 

# Daten verbinden
tab1_zg3 <- bind_cols(tab1_zg3_bas, tab1_zg3_kh)

tab1_alle <- tab1_zg3 %>% 
  mutate(oKH_tage= VZ_tage-KH_tage,
         VZ_tage_N = VZ_tage/N,
         KH_tage_N = KH_tage/N,
         oKH_tage_N = oKH_tage/N,
         per_KH_Tage= KH_tage/VZ_tage)

data1a <- data1 %>% 
  group_by(V_ID) %>% 
  arrange(START_DATE) %>% 
  mutate(n=row_number())

# nur erstes Intervall pro ID behalten
table(data1a$n)
data1b <- data1a %>% filter(n==1)

data1b$V_ALTER<- 
  factor(data1b$V_ALTER , 
         levels = c(3,4,5,6,7,8,9),
         labels = c("60-64","65-69","70-74","75-79","80-84","85-89","90 und mehr" ))

label(data1b$V_ALTER) <- "Altersgruppe"

data1b$tod<- 
  factor(data1b$tod , 
         levels=c(0,1),
         labels=c("nein", "ja"))

label(data1b$tod) <- "Tod"

label(data1b$V_PFLEGEGRAD) <- "Pflegegrad"

table1(~ V_ALTER + tod + V_PFLEGEGRAD| TG, data=data1b)
table1 <- table1(~ V_ALTER + tod + V_PFLEGEGRAD| TG, data=data1b)
table1 <- as.data.frame(table1)




### Pflegegrad am 30.06.2022 (ehemals 31.12.2022)
data1c <- data1 %>% 
  filter(END_DATE >= "2022-06-30") %>% 
  filter(V_STERBEDATUM >"2022-06-30"| is.na(V_STERBEDATUM))

data1c<- data1c %>% 
  group_by(V_ID) %>% 
  arrange(START_DATE) %>% 
  mutate(n=row_number())

# nur erstes Intervall pro ID behalten 
data1d <- data1c %>% filter(n==1)

data1d <- data1d %>% dplyr::select(V_ID,V_PFLEGEGRAD ) %>% 
  rename(PFLEGEGRAD_END = V_PFLEGEGRAD)

# mit vorherigem Datensatz verbinden 
data1e <- left_join(data1b, data1d, by=c("V_ID"))
data1f <- data1e %>% filter(!(is.na(PFLEGEGRAD_END))) %>% 
  mutate(V_PFLEGEGRAD2=as.numeric(V_PFLEGEGRAD),
         PFLEGEGRAD_END2=as.numeric(PFLEGEGRAD_END),
         diff_PG= PFLEGEGRAD_END2-V_PFLEGEGRAD2,
         diff_PG=as.factor(diff_PG))

table1_pg_mitte <- table1(~ V_PFLEGEGRAD + PFLEGEGRAD_END + diff_PG | TG, data=data1f)


### Pflegegrad am 31.12.2022
data1c <- data1 %>% 
  filter(END_DATE>="2022-12-31") %>% 
  filter(is.na(V_STERBEDATUM))

data1d<- data1c 

data1d <- data1d %>% 
  dplyr::select(V_ID,V_PFLEGEGRAD ) %>% 
  rename(PFLEGEGRAD_END = V_PFLEGEGRAD)

# mit vorherigem Datensatz verbinden 
data1e <- left_join(data1b, data1d, by=c("V_ID"))
data1f <- data1e %>% filter(!(is.na(PFLEGEGRAD_END))) %>% 
  mutate(V_PFLEGEGRAD2 = as.numeric(V_PFLEGEGRAD),
         PFLEGEGRAD_END2 =as.numeric(PFLEGEGRAD_END),
         diff_PG = PFLEGEGRAD_END2-V_PFLEGEGRAD2,
         diff_PG = as.factor(diff_PG))

table1_pg_ende <- table1(~ V_PFLEGEGRAD + PFLEGEGRAD_END + diff_PG | TG, data=data1f)

list_of_datasets <- list("table1" = table1,
                         "table1_pg_mitte" = table1_pg_mitte,
                         "table1_pg_ende" = table1_pg_ende)

write.xlsx(list_of_datasets , file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_Table1_2022.xlsx")
```

```{r descr_tg_2022, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(TG) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(TG) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("TG"))

tab1_tg <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```

```{r descr_pg_2022, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(V_PFLEGEGRAD) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(V_PFLEGEGRAD) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("V_PFLEGEGRAD"))

tab1_pg <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```


```{r descr_alter_2022, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(V_ALTER) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(V_ALTER) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("V_ALTER"))

tab1_age <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```


```{r descr_tod_2022, include = FALSE}
# ZG3 
tab1_zg3_bas <- data1 %>% group_by(tod) %>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  adorn_totals("row")

# ZG3-KH
tab1_zg3_kh <- data3a  %>%  group_by(tod) %>%
  summarise(KH_tage= sum(kh_dauer)) %>% 
  adorn_totals("row")

# Daten verbinden
tab1_zg3 <- left_join(tab1_zg3_bas, tab1_zg3_kh, by=c("tod"))

tab1_tod <- tab1_zg3 %>% mutate( oKH_tage= VZ_tage-KH_tage,
                                VZ_tage_N = VZ_tage/N,
                                KH_tage_N = KH_tage/N,
                                oKH_tage_N = oKH_tage/N,
                                per_KH_Tage= KH_tage/VZ_tage)
```

```{r export_2022}
list_of_datasets <- list("tab1_alle" = tab1_alle,
                         "tab1_tg" = tab1_tg,
                         "tab1_pg" = tab1_pg,
                         "tab1_age" = tab1_age,
                         "tab1_tod" = tab1_tod)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_Tab1_2022.xlsx")
```


## weitere Tabellen des Abschnitts
```{r tab1_VZ_Tage_2022}
# ZG3
tab1_zg3_VZ_Tage <- data1 %>% 
  filter(ZG3==1 ) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# ZG4
tab1_zg4_VZ_Tage <- data1 %>% 
  filter(ZG4 == 1 ) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG4", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# Daten verbinden
tab1_VZ_Tage <- bind_rows(tab1_zg3_VZ_Tage, tab1_zg4_VZ_Tage) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tab1_KH_Tage_2022}
# ZG3
tab1_zg3_KH_Tage <- data3a %>% 
  filter(ZG3==1 & !(is.na(kh_dauer))) %>%  
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# ZG4
tab1_zg4_KH_Tage <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>% 
  group_by(V_PFLEGEGRAD)%>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG4", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# Daten verbinden
tab1_KH_Tage <- bind_rows(tab1_zg3_KH_Tage, tab1_zg4_KH_Tage) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```


```{r tab1_oKH_Tage_2022}
tab1_oKH_Tage <- left_join(tab1_VZ_Tage, tab1_KH_Tage, by=c("V_PFLEGEGRAD","ZG")) %>% 
  mutate(n_oKH_Tage = n_Tage,
         oKH_Tage_voll =  VZ_Tage_voll-KH_Tage_voll,
         oKH_Tage = oKH_Tage_voll/n_oKH_Tage) 
```

```{r tab1_n_AG_2022}
n_AG_zg3 <- data1 %>% filter(ZG3 == 1) %>%
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row") %>% 
  mutate(ZG = "ZG3")

n_AG_zg4<- data1 %>% filter(ZG4 == 1) %>%  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE)%>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG4")

tab1_n_AG <- bind_rows(n_AG_zg3,n_AG_zg4) %>% relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tabelle5_2022}
# Data prep
data1 <- data1 %>% mutate(tod = if_else(is.na(V_STERBEDATUM),0,1))
data3a <- data3a %>% mutate(tod = if_else(is.na(V_STERBEDATUM),0,1))

# ZG3 
tab5_zg3_bas <- data1 %>% 
  filter(ZG3==1 ) %>%
  group_by(TG,V_PFLEGEGRAD, V_ALTER,tod,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer ))

# ZG3-KH
tab5_zg3_kh <- data3a %>%
  filter(ZG3==1 & !(is.na(kh_dauer)))%>%
  group_by(TG,V_PFLEGEGRAD, V_ALTER,tod,.drop = FALSE)%>% 
  summarise(KH_tage= sum(kh_dauer))  

# Daten verbinden
tab5_zg3 <- left_join(tab5_zg3_bas, tab5_zg3_kh, by = c("TG", "V_PFLEGEGRAD", "V_ALTER", "tod"))

tab5 <- tab5_zg3 %>% 
  filter(!(is.na(tod))) %>% 
  mutate( KH_tage = if_else(is.na(KH_tage), 0, KH_tage),
          oKH_Tage = VZ_tage - KH_tage,
          per_oKH_Tage = KH_tage / VZ_tage)

VZ_ZG_tab5 <- tab5
```

```{r tab5_VZ_Tage_2022}
# ZG3
tab5_zg3_VZ_Tage <- data1 %>% filter(ZG3==1 ) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG3", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# ZG4
tab5_zg4_VZ_Tage <- data1 %>% 
  filter(ZG3==1 ) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE)%>% 
  summarise(n_Tage = n_distinct(V_ID),
            VZ_Tage_voll = sum(intervalldauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG4", 
         VZ_Tage =  VZ_Tage_voll/n_Tage) 

# Daten verbinden
tab5_VZ_Tage <- bind_rows(tab5_zg3_VZ_Tage, tab5_zg4_VZ_Tage) %>% 
  relocate(ZG, .before = TG)
```


```{r tab5_KH_Tage_2022}
# ZG3
tab5_zg3_KH_Tage <- data3a %>% 
  filter(ZG3==1 & !(is.na(kh_dauer))) %>% 
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG3", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# ZG4
tab5_zg4_KH_Tage <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>%
  group_by(TG,V_PFLEGEGRAD,.drop = FALSE) %>% 
  summarise(n_KH_Tage = n_distinct(V_ID),
            KH_Tage_voll = sum(kh_dauer )) %>% 
  adorn_totals("row")%>% 
  mutate(ZG = "ZG4", 
         KH_Tage =  KH_Tage_voll/n_KH_Tage) 

# Daten verbinden
tab5_KH_Tage <- bind_rows(tab5_zg3_KH_Tage, tab5_zg4_KH_Tage) %>% relocate(ZG, .before = TG)
```

```{r tab5_oKH_Tage_2022}
tab5_oKH_Tage <- left_join(tab5_VZ_Tage,tab5_KH_Tage, by=c("TG", "V_PFLEGEGRAD","ZG")) %>% 
   mutate(n_oKH_Tage = n_Tage,
          oKH_Tage_voll =  VZ_Tage_voll-KH_Tage_voll,
          oKH_Tage = oKH_Tage_voll/n_oKH_Tage) 
```

```{r tab5_n_AG_2022}
n_AG_zg3 <- data1 %>% filter(ZG3==1 ) %>%  
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG3")

n_AG_zg4<- data1 %>% filter(ZG4==1 ) %>%  
  group_by(V_PFLEGEGRAD, V_ALTER, .drop=FALSE) %>% 
  summarise(n_AG = n_distinct(V_ID)) %>% 
  adorn_totals("row")%>% 
  mutate(ZG= "ZG4")

tab5_n_AG <- bind_rows(n_AG_zg3,n_AG_zg4) %>% 
  relocate(ZG, .before = V_PFLEGEGRAD)
```

```{r tabelle9_2022}
# Data prep
# ZG3 
tab9_zg3_bas <- data1 %>%
  filter(ZG3 == 1) %>%
  group_by(TG,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  mutate(ZG = "ZG3") 

# ZG4
tab9_zg4_bas <- data1 %>% 
  filter(ZG4 == 1) %>%  
  group_by(TG,.drop = FALSE)%>% 
  summarise(N = n_distinct(V_ID),
  VZ_tage = sum(intervalldauer )) %>% 
  mutate(ZG = "ZG4") 

# ZG3-KH
tab9_zg3_kh <- data3a %>% 
  filter(ZG3 == 1 & !(is.na(kh_dauer)))%>%
  group_by(TG,.drop = FALSE)%>% 
  summarise(KH_tage = sum(kh_dauer))  %>% 
  mutate(ZG = "ZG3") 

# ZG4-KH
tab9_zg4_kh <- data3a %>% 
  filter(ZG4==1 & !(is.na(kh_dauer))) %>% 
  group_by(TG,.drop = FALSE)%>% 
  summarise(KH_tage = sum(kh_dauer))  %>% 
  mutate(ZG = "ZG4") 

# Daten verbinden
tab9_zg3 <- left_join(tab9_zg3_bas, tab9_zg3_kh, by=c("TG","ZG"))
tab9_zg4 <- left_join(tab9_zg4_bas, tab9_zg4_kh, by=c("TG","ZG"))

tab9 <- bind_rows(tab9_zg3,tab9_zg4) %>% 
  mutate(oKH_Tage= VZ_tage-KH_tage,
         per_oKH_Tage= KH_tage/VZ_tage)

VZ_ZG_tab9 <- tab9 %>% relocate(ZG, .before = TG) 
```


```{r export2_2022}
# list_of_datasets <- list("VZ_ZG" = VZ_ZG,"tab1_VZ_Tage" =tab1_VZ_Tage,
#                          "tab1_KH_Tage" =tab1_KH_Tage,"tab1_oKH_Tage" =tab1_oKH_Tage,
#                          "tab1_n_AG" =tab1_n_AG)
# write.xlsx(list_of_datasets, 
#            file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/01_PG_2022.xlsx")


list_of_datasets <- list("VZ_ZG_tab5" = VZ_ZG_tab5,
                         "tab5_VZ_Tage" =tab5_VZ_Tage,
                         "tab5_KH_Tage" =tab5_KH_Tage,
                         "tab5_oKH_Tage" =tab5_oKH_Tage,
                         "tab5_n_AG" =tab5_n_AG)
write.xlsx(VZ_ZG_tab5, 
           file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/05_TG_PG_2022.xlsx")


list_of_datasets <- list("VZ_ZG_tab9" = VZ_ZG_tab9)
write.xlsx(list_of_datasets, 
           file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/01a_TG_2022.xlsx")
```
