==> R_Skripte/21_Inanspruch_Kosten_SGB5.Rmd <==
---
title: "Deskription Gruppen"
author: "ldp / grb"
output:
  html_document:
    toc: true
    toc_float:
      collapsed: false
---

```{r setup, include=FALSE}
library(tidyverse)
library(lubridate)
library(psych)
library(dlookr)
library(table1)
library(janitor)
library(openxlsx)
library(readxl)
#library(epiDisplay)
```


# 2019
```{r data_load_2019}
setwd("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/02_rdata/2019")

load("data11_neu.Rda")
data1 <- data11_neu

load("sgbv10.Rda")
load("kh11.Rda")

verweildauer_zg3_TG <- read_excel("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2019.xlsx",3)  

verweildauer_pg_TG_tod <- read_excel("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2019.xlsx",9) 
```


Vorbereitung KH-Daten
```{r kh_prep_2019, warning=FALSE}
### KH-Daten mit Stammdaten verbinden
# nur vollstationär behalten
kh11a<- kh11  %>%
  mutate(KH_AUF_ANLASS2 = str_pad(KH_AUF_ANLASS, 4, pad = "0"),
         KH_AUF_ANLASS_Beg = substr(KH_AUF_ANLASS2, 0, 2),
         KH_AUF_ANLASS_End = substr(KH_AUF_ANLASS2, 3, 4)) %>% 
  filter(KH_AUF_ANLASS_Beg == "01")

# überschneidende Hospitalisierungen verbinden
kh11a$KH_BEGINN_DAT <- ymd(kh11a$KH_BEGINN_DAT)
kh11a$KH_ENDE_DAT <- ymd(kh11a$KH_ENDE_DAT)
kh11a <- kh11a %>% 
  filter(!(is.na(KH_ENDE_DAT)))

kh11b<- kh11a%>%
  arrange(V_ID, KH_BEGINN_DAT) %>% 
  group_by(V_ID) %>%
  mutate(indx = c(0, cumsum(as.numeric(lead(KH_BEGINN_DAT )) > cummax(as.numeric(KH_ENDE_DAT )))[-n()])) %>%
  group_by(V_ID, indx) %>%
  summarise(KH_BEGINN_DAT2 = min(KH_BEGINN_DAT), 
            KH_ENDE_DAT2 = max(KH_ENDE_DAT))

#Klinikdaten für merge mit HM-Daten vorbereiten (rename und select)
kh11c <- kh11b %>% 
  mutate(LEISTUNG_VON = KH_BEGINN_DAT2,
         LEISTUNG_BIS = KH_ENDE_DAT2,
         origin = "KH") %>% 
  select(V_ID, LEISTUNG_VON, LEISTUNG_BIS, origin)
```


```{r, include = FALSE}
# HM-Daten vorbereiten
sgbv10$HM_DATUM <- ymd(sgbv10$HM_DATUM)
sgbv10$LEISTUNG_VON <- sgbv10$HM_DATUM
sgbv10$LEISTUNG_BIS <- sgbv10$HM_DATUM

# negative Kosten exkludieren
sgbv10 <- sgbv10 %>% filter(HM_GESKOSTEN>=0)

# HM-Datensatz für Verbindung erstellen
help1 <- sgbv10%>% arrange(V_ID) %>% 
  mutate(origin="HM")  

# Daten mit Hospitalisierungen (row_bind) und Stammdaten verbinden (filtern von fehlenden Beginn/Ende bzw. ohne Match)
help2 <- bind_rows(help1, kh11c)

data3 <- left_join(data1, help2, by = ("V_ID")) %>% 
  filter(!(is.na(LEISTUNG_VON))) %>% 
  filter(!(is.na(LEISTUNG_BIS))) 

# für jeden Tag eines Leistungserhaltes Zeile erstellen 
data3a <- data3 %>%
  # create sequence of days between start and end
  mutate(day = map2(LEISTUNG_VON, LEISTUNG_BIS, ~seq(.x, .y, "day"))) %>%
  # unnest data
  unnest(cols = c(day)) %>%                                                       
  group_by(V_ID,V_PFLEGEGRAD, TG,origin, day)

data3a2 <- data3a %>% 
  arrange(V_ID) %>% 
  ungroup() 

# nur relevante Tage  behalten (Filter vor allem notwendig wenn mehrere TGs und PG bei einer Person um passendes Intervall zu matchen, teilweise Leistungen nach Todesdatum)
data3b <- data3a2 %>%
  filter(START_DATE <= day & END_DATE >= day) %>% 
  mutate(HM = if_else(origin == "HM", 1, 0),
         kh = if_else(origin == "KH", 1, 0))

# für jede group_by Kombi prüfen ob an bestimmten Tag HM-leistung und/oder Hospitalisierung
data3c <- data3b %>% 
  group_by(V_ID, V_PFLEGEGRAD, TG, day) %>% 
  summarise(HM = max(HM),
            kh = max(kh))

# nur HM-Leistung ohne KH behalten
data3d <- data3c %>% 
  filter(!(kh == 1)) 

# HM-Leistungstage wieder mit Kosten verbinden
data3e <- left_join(data3d, sgbv10, by=c("V_ID", "day"="HM_DATUM")) %>% 
  select(-c(LEISTUNG_VON,LEISTUNG_BIS)) %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD)) 
```


Verbindung mit Gesamtdaten
```{r, include = FALSE}
data1 <- data1 %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD))

# Stammdaten mit Leistungstagen verbinden
data3f <- left_join(data1, data3e, by = c("V_ID", "V_PFLEGEGRAD", "TG"))

# nur Zeilen behalten die Stammdatenintervall und Leistungstag einschließen
data4 <- data3f %>% 
  filter(!(day > END_DATE | day < START_DATE)) %>% 
  rename(HM_DATUM = day)

data4 <- data4 %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD)) 

# unique Id pro Leistung
data4 <- data4 %>% mutate(id=row_number())
```


## Leistung nach TG
```{r}
# Data 1 vorbereiten
data1a_zg <- data1  %>%
  group_by(V_ID, TG) %>% 
  summarise(VZ_tage = sum(intervalldauer))

data1a_zg3 <- left_join(data1a_zg, verweildauer_zg3_TG, by=c("V_ID", "TG")) %>% 
  mutate(Verweildauer = if_else(is.na(Verweildauer),0,Verweildauer),
         oKH_Tage = VZ_tage- Verweildauer)

# Tage ohne KH_Aufenthalt
data1a_zg3_verweil <- data1a_zg3 %>% 
  group_by( TG, .drop=FALSE) %>% 
  summarise( VZ_tage = sum(VZ_tage),
             oKH_Tage = sum(oKH_Tage))


### auf ID-Basis
data4_zg3_id <- data4  %>%  
  group_by(V_ID,TG) %>% 
  summarise(HM_ja_id = 1) %>% 
  group_by(TG) %>% 
  summarise(n_IN_HM = sum(HM_ja_id))

data4_zg3_id2 <- data1  %>%  
  group_by(V_ID,TG) %>%
  summarise(id_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(n_ID = sum(id_ja))
  
data4_zg3_id3 <- left_join(data4_zg3_id2,data4_zg3_id, by=c("TG")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(n_noIN_HM = n_ID- n_IN_HM)

# Gesamttage mit HM-Leistung und Tage ohne HM-Leistung 
data4_zg3_tage1 <- data4 %>%
  group_by(V_ID,TG, HM_DATUM) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(IN_Tage_HM = sum(ID_ja))

data4_zg3_tage2<-  left_join(data1a_zg3_verweil,data4_zg3_tage1, by=c("TG")  ) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(noIN_Tage_HM= oKH_Tage -IN_Tage_HM) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtleistungen HM pro TG 
data4_zg3_leist1 <- data4 %>%  
  group_by(V_ID,TG,id) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(Leistungen_HM = sum(ID_ja))

data4_zg3_leist2<-  left_join(data1a_zg3_verweil,data4_zg3_leist1, by=c("TG")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))


### Gesamtkosten, Zuzahlung, Eigenanteil, Mehrkosten
data4_zg3_kosten <- data4  %>% group_by(TG, .drop=FALSE) %>%
  summarise(SUM_HM_GesKosten = sum(HM_GESKOSTEN),
            SUM_HM_Zuzahlung = sum(HM_ZUZAHLUNG),
            SUM_HM_Eigenanteil = sum(HM_EIGENANTEIL),
            SUM_HM_Mehrkosten = sum(HM_MEHRKOSTEN))


# alles verbinden
data4_zg3_a <- left_join(data1a_zg3_verweil, data4_zg3_id3, by=c("TG"))
data4_zg3_b <- left_join(data4_zg3_a,data4_zg3_tage2, by=c("TG"))
data4_zg3_c <- left_join(data4_zg3_b, data4_zg3_leist2, by=c("TG")) 
data4_zg3_d <- left_join(data4_zg3_c, data4_zg3_kosten, by=c("TG")) 

# Daten für Tabelle vorbereiten
results_alle <- data4_zg3_d %>% 
  mutate(proz_n_IN_HM = n_IN_HM/n_ID,
         Kosten_Leistung = SUM_HM_GesKosten/Leistungen_HM,
         Leistungen_HM_jahr = Leistungen_HM*365/oKH_Tage,
         HM_GesKosten_jahr = SUM_HM_GesKosten*365/oKH_Tage)

# relevante Variablen für Tabelle
results_alle <- results_alle %>% 
  dplyr::select(TG, n_ID, VZ_tage,oKH_Tage,n_IN_HM,Leistungen_HM,
                proz_n_IN_HM,Kosten_Leistung,Leistungen_HM_jahr,HM_GesKosten_jahr) %>% 
  mutate_all(~replace(., is.na(.), 0))  
```

## Leistung nach ZG, PG und TG
```{r}
# Data 1 vorbereiten
data1a_zg <- data1  %>%  
  group_by(V_ID, V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(VZ_tage = sum(intervalldauer ))

verweildauer_pg_TG_tod$V_ALTER <- as.integer(verweildauer_pg_TG_tod$V_ALTER )
data1a_zg3 <- left_join(data1a_zg, verweildauer_pg_TG_tod, by=c("V_ID", "V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate(Verweildauer = if_else(is.na(Verweildauer), 0, Verweildauer),
         oKH_Tage = VZ_tage - Verweildauer )

# Tage ohne KH_Aufenthalt 
data1a_zg3_verweil <- data1a_zg3 %>% 
  group_by( V_PFLEGEGRAD,TG,V_ALTER, tod, .drop=FALSE) %>% 
  summarise(VZ_tage = sum(VZ_tage),
            oKH_Tage = sum(oKH_Tage))

# auf ID-Basis 
data4_zg3_id <- data4  %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(HM_ja_id = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(n_IN_HM = sum(HM_ja_id))

data4_zg3_id2 <- data1  %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod) %>%
  summarise(id_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(n_ID= sum(id_ja))
  
data4_zg3_id3 <- left_join(data4_zg3_id2,data4_zg3_id, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod") ) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(n_noIN_HM=n_ID- n_IN_HM)

# Gesamttage mit HM-Leistung und Tage ohne HM-Leistung 
data4_zg3_tage1 <- data4 %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod, HM_DATUM) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(IN_Tage_HM = sum(ID_ja) )

data4_zg3_tage2<-  left_join(data1a_zg3_verweil,data4_zg3_tage1,by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(noIN_Tage_HM = oKH_Tage -IN_Tage_HM) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtleistungen HM pro TG 
data4_zg3_leist1 <- data4 %>%
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod,id) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(Leistungen_HM = sum(ID_ja))

data4_zg3_leist2<-  left_join(data1a_zg3_verweil,data4_zg3_leist1, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0))  %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtkosten, Zuzahlung, Eigenanteil, Mehrkosten
data4_zg3_kosten <- data4  %>% 
  group_by( V_PFLEGEGRAD,TG,V_ALTER, tod) %>%
  summarise(SUM_HM_GesKosten = sum(HM_GESKOSTEN),
            SUM_HM_Zuzahlung = sum(HM_ZUZAHLUNG),
            SUM_HM_Eigenanteil = sum(HM_EIGENANTEIL),
            SUM_HM_Mehrkosten = sum(HM_MEHRKOSTEN))

# alles verbinden
data4_zg3_a <- left_join(data1a_zg3_verweil, data4_zg3_id3, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))
data4_zg3_b <- left_join(data4_zg3_a,data4_zg3_tage2, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0)) 
data4_zg3_c <- left_join(data4_zg3_b, data4_zg3_leist2, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))
data4_zg3_d <- left_join(data4_zg3_c, data4_zg3_kosten,by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))

# Daten für Tabelle vorbereiten
results_alle_w  <- data4_zg3_d %>% 
  mutate(proz_n_IN_HM = n_IN_HM/n_ID,
         Kosten_Leistung = SUM_HM_GesKosten/Leistungen_HM,
         Leistungen_HM_jahr = Leistungen_HM*365/oKH_Tage,
         HM_GesKosten_jahr = SUM_HM_GesKosten*365/oKH_Tage)

# relevante Variablen für Tabelle
results_alle_w <- results_alle_w %>% 
  dplyr::select(TG,V_PFLEGEGRAD,V_ALTER, tod, n_ID, VZ_tage,oKH_Tage,n_IN_HM,Leistungen_HM,Kosten_Leistung, proz_n_IN_HM,Leistungen_HM_jahr,HM_GesKosten_jahr) %>% 
  mutate_all(~replace(., is.na(.), 0))  
```

```{r wichtung_2019}
# zunächst Anzahl der VZ_TAGE pro TG
data5 <- results_alle_w %>%  
  group_by(TG) %>% 
  summarise(days_tg = sum(oKH_Tage))

# mit Ergebnise verbinden 
data5a <- left_join(results_alle_w, data5, by=c("TG"))

# Wichtung nach TG1
data6 <- data5a %>% 
  filter(TG=="TG1") %>% 
  mutate(weight=oKH_Tage/days_tg) %>% 
  ungroup() %>% 
  dplyr::select(c("V_PFLEGEGRAD","V_ALTER", "tod","weight"))

# erneute Verbindung mit Ergebnissen
data6a <- left_join(data5a, data6, by=c("V_PFLEGEGRAD","V_ALTER", "tod")) %>%
  mutate_all(~replace(., is.na(.), 0))

# Wichtung anwenden
data6b <- data6a %>% mutate(Leistungen_HM_jahr_w = Leistungen_HM_jahr * weight,
                            HM_GesKosten_jahr_w = HM_GesKosten_jahr * weight)


# finale Werte
data7 <- data6b %>% 
  group_by(TG) %>% 
  filter(weight > 0) %>% 
  summarise(oKH_Tage = sum(oKH_Tage),
            Leistungen_HM_jahr = sum(Leistungen_HM_jahr_w),
            HM_GesKosten_jahr = sum(HM_GesKosten_jahr_w))

# for n
help1 <- left_join(data1,data6, by=c("V_PFLEGEGRAD","V_ALTER", "tod")) %>% 
  filter(!(is.na(weight)))

n_weight <- help1 %>% 
  group_by(TG) %>% 
  summarise(N=n_distinct(V_ID))

# zusammenführen
results_weight <- left_join(n_weight, data7, by=c("TG"))
```

```{r export_2019}
list_of_datasets <- list("results_alle" = results_alle, "results_weight" = results_weight)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/21_Kosten_SGB5_2019.xlsx")
```



# 2020
```{r data_load_2020}
setwd("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/02_rdata/2020")

load("data11_neu.Rda")
data1 <- data11_neu

load("sgbv10.Rda")
load("kh11.Rda")

verweildauer_zg3_TG <- read_excel("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2020.xlsx",3)  

verweildauer_pg_TG_tod <- read_excel("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2020.xlsx",9) 
```


Vorbereitung KH-Daten
```{r kh_prep_2020, warning=FALSE}
### KH-Daten mit Stammdaten verbinden
# nur vollstationär behalten
kh11a<- kh11  %>%
  mutate(KH_AUF_ANLASS2 = str_pad(KH_AUF_ANLASS, 4, pad = "0"),
         KH_AUF_ANLASS_Beg = substr(KH_AUF_ANLASS2, 0, 2),
         KH_AUF_ANLASS_End = substr(KH_AUF_ANLASS2, 3, 4)) %>% 
  filter(KH_AUF_ANLASS_Beg == "01")

# überschneidende Hospitalisierungen verbinden
kh11a$KH_BEGINN_DAT <- ymd(kh11a$KH_BEGINN_DAT)
kh11a$KH_ENDE_DAT <- ymd(kh11a$KH_ENDE_DAT)
kh11a <- kh11a %>% 
  filter(!(is.na(KH_ENDE_DAT)))

kh11b<- kh11a%>%
  arrange(V_ID, KH_BEGINN_DAT) %>% 
  group_by(V_ID) %>%
  mutate(indx = c(0, cumsum(as.numeric(lead(KH_BEGINN_DAT )) > cummax(as.numeric(KH_ENDE_DAT )))[-n()])) %>%
  group_by(V_ID, indx) %>%
  summarise(KH_BEGINN_DAT2 = min(KH_BEGINN_DAT), 
            KH_ENDE_DAT2 = max(KH_ENDE_DAT))

#Klinikdaten für merge mit HM-Daten vorbereiten (rename und select)
kh11c <- kh11b %>% 
  mutate(LEISTUNG_VON = KH_BEGINN_DAT2,
         LEISTUNG_BIS = KH_ENDE_DAT2,
         origin = "KH") %>% 
  select(V_ID, LEISTUNG_VON, LEISTUNG_BIS, origin)
```


```{r, include = FALSE}
# HM-Daten vorbereiten
sgbv10$HM_DATUM <- ymd(sgbv10$HM_DATUM)
sgbv10$LEISTUNG_VON <- sgbv10$HM_DATUM
sgbv10$LEISTUNG_BIS <- sgbv10$HM_DATUM

# negative Kosten exkludieren
sgbv10 <- sgbv10 %>% filter(HM_GESKOSTEN>=0)

# HM-Datensatz für Verbindung erstellen
help1 <- sgbv10%>% arrange(V_ID) %>% 
  mutate(origin="HM")  

# Daten mit Hospitalisierungen (row_bind) und Stammdaten verbinden (filtern von fehlenden Beginn/Ende bzw. ohne Match)
help2 <- bind_rows(help1, kh11c)

data3 <- left_join(data1, help2, by = ("V_ID")) %>% 
  filter(!(is.na(LEISTUNG_VON))) %>% 
  filter(!(is.na(LEISTUNG_BIS))) 

# für jeden Tag eines Leistungserhaltes Zeile erstellen 
data3a <- data3 %>%
  # create sequence of days between start and end
  mutate(day = map2(LEISTUNG_VON, LEISTUNG_BIS, ~seq(.x, .y, "day"))) %>%
  # unnest data
  unnest(cols = c(day)) %>%                                                       
  group_by(V_ID,V_PFLEGEGRAD, TG,origin, day)

data3a2 <- data3a %>% 
  arrange(V_ID) %>% 
  ungroup() 

# nur relevante Tage  behalten (Filter vor allem notwendig wenn mehrere TGs und PG bei einer Person um passendes Intervall zu matchen, teilweise Leistungen nach Todesdatum)
data3b <- data3a2 %>%
  filter(START_DATE <= day & END_DATE >= day) %>% 
  mutate(HM = if_else(origin == "HM", 1, 0),
         kh = if_else(origin == "KH", 1, 0))

# für jede group_by Kombi prüfen ob an bestimmten Tag HM-leistung und/oder Hospitalisierung
data3c <- data3b %>% 
  group_by(V_ID, V_PFLEGEGRAD, TG, day) %>% 
  summarise(HM = max(HM),
            kh = max(kh))

# nur HM-Leistung ohne KH behalten
data3d <- data3c %>% 
  filter(!(kh == 1)) 

# HM-Leistungstage wieder mit Kosten verbinden
data3e <- left_join(data3d, sgbv10, by=c("V_ID", "day"="HM_DATUM")) %>% 
  select(-c(LEISTUNG_VON,LEISTUNG_BIS)) %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD)) 
```


Verbindung mit Gesamtdaten
```{r, include = FALSE}
data1 <- data1 %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD))

# Stammdaten mit Leistungstagen verbinden
data3f <- left_join(data1, data3e, by = c("V_ID", "V_PFLEGEGRAD", "TG"))

# nur Zeilen behalten die Stammdatenintervall und Leistungstag einschließen
data4 <- data3f %>% 
  filter(!(day > END_DATE | day < START_DATE)) %>% 
  rename(HM_DATUM = day)

data4 <- data4 %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD)) 

# unique Id pro Leistung
data4 <- data4 %>% mutate(id=row_number())
```


## Leistung nach TG
```{r}
# Data 1 vorbereiten
data1a_zg <- data1  %>%
  group_by(V_ID, TG) %>% 
  summarise(VZ_tage = sum(intervalldauer))

data1a_zg3 <- left_join(data1a_zg, verweildauer_zg3_TG, by=c("V_ID", "TG")) %>% 
  mutate(Verweildauer = if_else(is.na(Verweildauer),0,Verweildauer),
         oKH_Tage = VZ_tage- Verweildauer)

# Tage ohne KH_Aufenthalt
data1a_zg3_verweil <- data1a_zg3 %>% 
  group_by( TG, .drop=FALSE) %>% 
  summarise( VZ_tage = sum(VZ_tage),
             oKH_Tage = sum(oKH_Tage))


### auf ID-Basis
data4_zg3_id <- data4  %>%  
  group_by(V_ID,TG) %>% 
  summarise(HM_ja_id = 1) %>% 
  group_by(TG) %>% 
  summarise(n_IN_HM = sum(HM_ja_id))

data4_zg3_id2 <- data1  %>%  
  group_by(V_ID,TG) %>%
  summarise(id_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(n_ID = sum(id_ja))
  
data4_zg3_id3 <- left_join(data4_zg3_id2,data4_zg3_id, by=c("TG")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(n_noIN_HM = n_ID- n_IN_HM)

# Gesamttage mit HM-Leistung und Tage ohne HM-Leistung 
data4_zg3_tage1 <- data4 %>%
  group_by(V_ID,TG, HM_DATUM) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(IN_Tage_HM = sum(ID_ja))

data4_zg3_tage2<-  left_join(data1a_zg3_verweil,data4_zg3_tage1, by=c("TG")  ) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(noIN_Tage_HM= oKH_Tage -IN_Tage_HM) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtleistungen HM pro TG 
data4_zg3_leist1 <- data4 %>%  
  group_by(V_ID,TG,id) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(Leistungen_HM = sum(ID_ja))

data4_zg3_leist2<-  left_join(data1a_zg3_verweil,data4_zg3_leist1, by=c("TG")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))


### Gesamtkosten, Zuzahlung, Eigenanteil, Mehrkosten
data4_zg3_kosten <- data4  %>% group_by(TG, .drop=FALSE) %>%
  summarise(SUM_HM_GesKosten = sum(HM_GESKOSTEN),
            SUM_HM_Zuzahlung = sum(HM_ZUZAHLUNG),
            SUM_HM_Eigenanteil = sum(HM_EIGENANTEIL),
            SUM_HM_Mehrkosten = sum(HM_MEHRKOSTEN))


# alles verbinden
data4_zg3_a <- left_join(data1a_zg3_verweil, data4_zg3_id3, by=c("TG"))
data4_zg3_b <- left_join(data4_zg3_a,data4_zg3_tage2, by=c("TG"))
data4_zg3_c <- left_join(data4_zg3_b, data4_zg3_leist2, by=c("TG")) 
data4_zg3_d <- left_join(data4_zg3_c, data4_zg3_kosten, by=c("TG")) 

# Daten für Tabelle vorbereiten
results_alle <- data4_zg3_d %>% 
  mutate(proz_n_IN_HM = n_IN_HM/n_ID,
         Kosten_Leistung = SUM_HM_GesKosten/Leistungen_HM,
         Leistungen_HM_jahr = Leistungen_HM*365/oKH_Tage,
         HM_GesKosten_jahr = SUM_HM_GesKosten*365/oKH_Tage)

# relevante Variablen für Tabelle
results_alle <- results_alle %>% 
  dplyr::select(TG, n_ID, VZ_tage,oKH_Tage,n_IN_HM,Leistungen_HM,
                proz_n_IN_HM,Kosten_Leistung,Leistungen_HM_jahr,HM_GesKosten_jahr) %>% 
  mutate_all(~replace(., is.na(.), 0))  
```

## Leistung nach ZG, PG und TG
```{r}
# Data 1 vorbereiten
data1a_zg <- data1  %>%  
  group_by(V_ID, V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(VZ_tage = sum(intervalldauer ))

verweildauer_pg_TG_tod$V_ALTER <- as.integer(verweildauer_pg_TG_tod$V_ALTER )
data1a_zg3 <- left_join(data1a_zg, verweildauer_pg_TG_tod, by=c("V_ID", "V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate(Verweildauer = if_else(is.na(Verweildauer), 0, Verweildauer),
         oKH_Tage = VZ_tage - Verweildauer )

# Tage ohne KH_Aufenthalt 
data1a_zg3_verweil <- data1a_zg3 %>% 
  group_by( V_PFLEGEGRAD,TG,V_ALTER, tod, .drop=FALSE) %>% 
  summarise(VZ_tage = sum(VZ_tage),
            oKH_Tage = sum(oKH_Tage))

# auf ID-Basis 
data4_zg3_id <- data4  %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(HM_ja_id = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(n_IN_HM = sum(HM_ja_id))

data4_zg3_id2 <- data1  %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod) %>%
  summarise(id_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(n_ID= sum(id_ja))
  
data4_zg3_id3 <- left_join(data4_zg3_id2,data4_zg3_id, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod") ) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(n_noIN_HM=n_ID- n_IN_HM)

# Gesamttage mit HM-Leistung und Tage ohne HM-Leistung 
data4_zg3_tage1 <- data4 %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod, HM_DATUM) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(IN_Tage_HM = sum(ID_ja) )

data4_zg3_tage2<-  left_join(data1a_zg3_verweil,data4_zg3_tage1,by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(noIN_Tage_HM = oKH_Tage -IN_Tage_HM) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtleistungen HM pro TG 
data4_zg3_leist1 <- data4 %>%
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod,id) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(Leistungen_HM = sum(ID_ja))

data4_zg3_leist2<-  left_join(data1a_zg3_verweil,data4_zg3_leist1, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0))  %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtkosten, Zuzahlung, Eigenanteil, Mehrkosten
data4_zg3_kosten <- data4  %>% 
  group_by( V_PFLEGEGRAD,TG,V_ALTER, tod) %>%
  summarise(SUM_HM_GesKosten = sum(HM_GESKOSTEN),
            SUM_HM_Zuzahlung = sum(HM_ZUZAHLUNG),
            SUM_HM_Eigenanteil = sum(HM_EIGENANTEIL),
            SUM_HM_Mehrkosten = sum(HM_MEHRKOSTEN))

# alles verbinden
data4_zg3_a <- left_join(data1a_zg3_verweil, data4_zg3_id3, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))
data4_zg3_b <- left_join(data4_zg3_a,data4_zg3_tage2, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0)) 
data4_zg3_c <- left_join(data4_zg3_b, data4_zg3_leist2, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))
data4_zg3_d <- left_join(data4_zg3_c, data4_zg3_kosten,by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))

# Daten für Tabelle vorbereiten
results_alle_w  <- data4_zg3_d %>% 
  mutate(proz_n_IN_HM = n_IN_HM/n_ID,
         Kosten_Leistung = SUM_HM_GesKosten/Leistungen_HM,
         Leistungen_HM_jahr = Leistungen_HM*365/oKH_Tage,
         HM_GesKosten_jahr = SUM_HM_GesKosten*365/oKH_Tage)

# relevante Variablen für Tabelle
results_alle_w <- results_alle_w %>% 
  dplyr::select(TG,V_PFLEGEGRAD,V_ALTER, tod, n_ID, VZ_tage,oKH_Tage,n_IN_HM,Leistungen_HM,Kosten_Leistung, proz_n_IN_HM,Leistungen_HM_jahr,HM_GesKosten_jahr) %>% 
  mutate_all(~replace(., is.na(.), 0))  
```

```{r wichtung_2020}
# zunächst Anzahl der VZ_TAGE pro TG
data5 <- results_alle_w %>%  
  group_by(TG) %>% 
  summarise(days_tg = sum(oKH_Tage))

# mit Ergebnise verbinden 
data5a <- left_join(results_alle_w, data5, by=c("TG"))

# Wichtung nach TG1
data6 <- data5a %>% 
  filter(TG=="TG1") %>% 
  mutate(weight=oKH_Tage/days_tg) %>% 
  ungroup() %>% 
  dplyr::select(c("V_PFLEGEGRAD","V_ALTER", "tod","weight"))

# erneute Verbindung mit Ergebnissen
data6a <- left_join(data5a, data6, by=c("V_PFLEGEGRAD","V_ALTER", "tod")) %>%
  mutate_all(~replace(., is.na(.), 0))

# Wichtung anwenden
data6b <- data6a %>% mutate(Leistungen_HM_jahr_w = Leistungen_HM_jahr * weight,
                            HM_GesKosten_jahr_w = HM_GesKosten_jahr * weight)


# finale Werte
data7 <- data6b %>% 
  group_by(TG) %>% 
  filter(weight > 0) %>% 
  summarise(oKH_Tage = sum(oKH_Tage),
            Leistungen_HM_jahr = sum(Leistungen_HM_jahr_w),
            HM_GesKosten_jahr = sum(HM_GesKosten_jahr_w))

# for n
help1 <- left_join(data1,data6, by=c("V_PFLEGEGRAD","V_ALTER", "tod")) %>% 
  filter(!(is.na(weight)))

n_weight <- help1 %>% 
  group_by(TG) %>% 
  summarise(N=n_distinct(V_ID))

# zusammenführen
results_weight <- left_join(n_weight, data7, by=c("TG"))
```

```{r export_2020}
list_of_datasets <- list("results_alle" = results_alle, "results_weight" = results_weight)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/21_Kosten_SGB5_2020.xlsx")
```


# 2021
```{r data_load_2021}
setwd("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/02_rdata/2021")

load("data11_neu.Rda")
data1 <- data11_neu

load("sgbv10.Rda")
load("kh11.Rda")

verweildauer_zg3_TG <- read_excel("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2021.xlsx",3)  

verweildauer_pg_TG_tod <- read_excel("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2021.xlsx",9) 
```


Vorbereitung KH-Daten
```{r kh_prep_2021, warning=FALSE}
### KH-Daten mit Stammdaten verbinden
# nur vollstationär behalten
kh11a<- kh11  %>%
  mutate(KH_AUF_ANLASS2 = str_pad(KH_AUF_ANLASS, 4, pad = "0"),
         KH_AUF_ANLASS_Beg = substr(KH_AUF_ANLASS2, 0, 2),
         KH_AUF_ANLASS_End = substr(KH_AUF_ANLASS2, 3, 4)) %>% 
  filter(KH_AUF_ANLASS_Beg == "01")

# überschneidende Hospitalisierungen verbinden
kh11a$KH_BEGINN_DAT <- ymd(kh11a$KH_BEGINN_DAT)
kh11a$KH_ENDE_DAT <- ymd(kh11a$KH_ENDE_DAT)
kh11a <- kh11a %>% 
  filter(!(is.na(KH_ENDE_DAT)))

kh11b<- kh11a%>%
  arrange(V_ID, KH_BEGINN_DAT) %>% 
  group_by(V_ID) %>%
  mutate(indx = c(0, cumsum(as.numeric(lead(KH_BEGINN_DAT )) > cummax(as.numeric(KH_ENDE_DAT )))[-n()])) %>%
  group_by(V_ID, indx) %>%
  summarise(KH_BEGINN_DAT2 = min(KH_BEGINN_DAT), 
            KH_ENDE_DAT2 = max(KH_ENDE_DAT))

#Klinikdaten für merge mit HM-Daten vorbereiten (rename und select)
kh11c <- kh11b %>% 
  mutate(LEISTUNG_VON = KH_BEGINN_DAT2,
         LEISTUNG_BIS = KH_ENDE_DAT2,
         origin = "KH") %>% 
  select(V_ID, LEISTUNG_VON, LEISTUNG_BIS, origin)
```


```{r, include = FALSE}
# HM-Daten vorbereiten
sgbv10$HM_DATUM <- ymd(sgbv10$HM_DATUM)
sgbv10$LEISTUNG_VON <- sgbv10$HM_DATUM
sgbv10$LEISTUNG_BIS <- sgbv10$HM_DATUM

# negative Kosten exkludieren
sgbv10 <- sgbv10 %>% filter(HM_GESKOSTEN>=0)

# HM-Datensatz für Verbindung erstellen
help1 <- sgbv10%>% arrange(V_ID) %>% 
  mutate(origin="HM")  

# Daten mit Hospitalisierungen (row_bind) und Stammdaten verbinden (filtern von fehlenden Beginn/Ende bzw. ohne Match)
help2 <- bind_rows(help1, kh11c)

data3 <- left_join(data1, help2, by = ("V_ID")) %>% 
  filter(!(is.na(LEISTUNG_VON))) %>% 
  filter(!(is.na(LEISTUNG_BIS))) 

# für jeden Tag eines Leistungserhaltes Zeile erstellen 
data3a <- data3 %>%
  # create sequence of days between start and end
  mutate(day = map2(LEISTUNG_VON, LEISTUNG_BIS, ~seq(.x, .y, "day"))) %>%
  # unnest data
  unnest(cols = c(day)) %>%                                                       
  group_by(V_ID,V_PFLEGEGRAD, TG,origin, day)

data3a2 <- data3a %>% 
  arrange(V_ID) %>% 
  ungroup() 

# nur relevante Tage  behalten (Filter vor allem notwendig wenn mehrere TGs und PG bei einer Person um passendes Intervall zu matchen, teilweise Leistungen nach Todesdatum)
data3b <- data3a2 %>%
  filter(START_DATE <= day & END_DATE >= day) %>% 
  mutate(HM = if_else(origin == "HM", 1, 0),
         kh = if_else(origin == "KH", 1, 0))

# für jede group_by Kombi prüfen ob an bestimmten Tag HM-leistung und/oder Hospitalisierung
data3c <- data3b %>% 
  group_by(V_ID, V_PFLEGEGRAD, TG, day) %>% 
  summarise(HM = max(HM),
            kh = max(kh))

# nur HM-Leistung ohne KH behalten
data3d <- data3c %>% 
  filter(!(kh == 1)) 

# HM-Leistungstage wieder mit Kosten verbinden
data3e <- left_join(data3d, sgbv10, by=c("V_ID", "day"="HM_DATUM")) %>% 
  select(-c(LEISTUNG_VON,LEISTUNG_BIS)) %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD)) 
```


Verbindung mit Gesamtdaten
```{r, include = FALSE}
data1 <- data1 %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD))

# Stammdaten mit Leistungstagen verbinden
data3f <- left_join(data1, data3e, by = c("V_ID", "V_PFLEGEGRAD", "TG"))

# nur Zeilen behalten die Stammdatenintervall und Leistungstag einschließen
data4 <- data3f %>% 
  filter(!(day > END_DATE | day < START_DATE)) %>% 
  rename(HM_DATUM = day)

data4 <- data4 %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD)) 

# unique Id pro Leistung
data4 <- data4 %>% mutate(id=row_number())
```


## Leistung nach TG
```{r}
# Data 1 vorbereiten
data1a_zg <- data1  %>%
  group_by(V_ID, TG) %>% 
  summarise(VZ_tage = sum(intervalldauer))

data1a_zg3 <- left_join(data1a_zg, verweildauer_zg3_TG, by=c("V_ID", "TG")) %>% 
  mutate(Verweildauer = if_else(is.na(Verweildauer),0,Verweildauer),
         oKH_Tage = VZ_tage- Verweildauer)

# Tage ohne KH_Aufenthalt
data1a_zg3_verweil <- data1a_zg3 %>% 
  group_by( TG, .drop=FALSE) %>% 
  summarise( VZ_tage = sum(VZ_tage),
             oKH_Tage = sum(oKH_Tage))


### auf ID-Basis
data4_zg3_id <- data4  %>%  
  group_by(V_ID,TG) %>% 
  summarise(HM_ja_id = 1) %>% 
  group_by(TG) %>% 
  summarise(n_IN_HM = sum(HM_ja_id))

data4_zg3_id2 <- data1  %>%  
  group_by(V_ID,TG) %>%
  summarise(id_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(n_ID = sum(id_ja))
  
data4_zg3_id3 <- left_join(data4_zg3_id2,data4_zg3_id, by=c("TG")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(n_noIN_HM = n_ID- n_IN_HM)

# Gesamttage mit HM-Leistung und Tage ohne HM-Leistung 
data4_zg3_tage1 <- data4 %>%
  group_by(V_ID,TG, HM_DATUM) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(IN_Tage_HM = sum(ID_ja))

data4_zg3_tage2<-  left_join(data1a_zg3_verweil,data4_zg3_tage1, by=c("TG")  ) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(noIN_Tage_HM= oKH_Tage -IN_Tage_HM) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtleistungen HM pro TG 
data4_zg3_leist1 <- data4 %>%  
  group_by(V_ID,TG,id) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(Leistungen_HM = sum(ID_ja))

data4_zg3_leist2<-  left_join(data1a_zg3_verweil,data4_zg3_leist1, by=c("TG")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))


### Gesamtkosten, Zuzahlung, Eigenanteil, Mehrkosten
data4_zg3_kosten <- data4  %>% group_by(TG, .drop=FALSE) %>%
  summarise(SUM_HM_GesKosten = sum(HM_GESKOSTEN),
            SUM_HM_Zuzahlung = sum(HM_ZUZAHLUNG),
            SUM_HM_Eigenanteil = sum(HM_EIGENANTEIL),
            SUM_HM_Mehrkosten = sum(HM_MEHRKOSTEN))


# alles verbinden
data4_zg3_a <- left_join(data1a_zg3_verweil, data4_zg3_id3, by=c("TG"))
data4_zg3_b <- left_join(data4_zg3_a,data4_zg3_tage2, by=c("TG"))
data4_zg3_c <- left_join(data4_zg3_b, data4_zg3_leist2, by=c("TG")) 
data4_zg3_d <- left_join(data4_zg3_c, data4_zg3_kosten, by=c("TG")) 

# Daten für Tabelle vorbereiten
results_alle <- data4_zg3_d %>% 
  mutate(proz_n_IN_HM = n_IN_HM/n_ID,
         Kosten_Leistung = SUM_HM_GesKosten/Leistungen_HM,
         Leistungen_HM_jahr = Leistungen_HM*365/oKH_Tage,
         HM_GesKosten_jahr = SUM_HM_GesKosten*365/oKH_Tage)

# relevante Variablen für Tabelle
results_alle <- results_alle %>% 
  dplyr::select(TG, n_ID, VZ_tage,oKH_Tage,n_IN_HM,Leistungen_HM,
                proz_n_IN_HM,Kosten_Leistung,Leistungen_HM_jahr,HM_GesKosten_jahr) %>% 
  mutate_all(~replace(., is.na(.), 0))  
```

## Leistung nach ZG, PG und TG
```{r}
# Data 1 vorbereiten
data1a_zg <- data1  %>%  
  group_by(V_ID, V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(VZ_tage = sum(intervalldauer ))

verweildauer_pg_TG_tod$V_ALTER <- as.integer(verweildauer_pg_TG_tod$V_ALTER )
data1a_zg3 <- left_join(data1a_zg, verweildauer_pg_TG_tod, by=c("V_ID", "V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate(Verweildauer = if_else(is.na(Verweildauer), 0, Verweildauer),
         oKH_Tage = VZ_tage - Verweildauer )

# Tage ohne KH_Aufenthalt 
data1a_zg3_verweil <- data1a_zg3 %>% 
  group_by( V_PFLEGEGRAD,TG,V_ALTER, tod, .drop=FALSE) %>% 
  summarise(VZ_tage = sum(VZ_tage),
            oKH_Tage = sum(oKH_Tage))

# auf ID-Basis 
data4_zg3_id <- data4  %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(HM_ja_id = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(n_IN_HM = sum(HM_ja_id))

data4_zg3_id2 <- data1  %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod) %>%
  summarise(id_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(n_ID= sum(id_ja))
  
data4_zg3_id3 <- left_join(data4_zg3_id2,data4_zg3_id, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod") ) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(n_noIN_HM=n_ID- n_IN_HM)

# Gesamttage mit HM-Leistung und Tage ohne HM-Leistung 
data4_zg3_tage1 <- data4 %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod, HM_DATUM) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(IN_Tage_HM = sum(ID_ja) )

data4_zg3_tage2<-  left_join(data1a_zg3_verweil,data4_zg3_tage1,by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(noIN_Tage_HM = oKH_Tage -IN_Tage_HM) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtleistungen HM pro TG 
data4_zg3_leist1 <- data4 %>%
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod,id) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(Leistungen_HM = sum(ID_ja))

data4_zg3_leist2<-  left_join(data1a_zg3_verweil,data4_zg3_leist1, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0))  %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtkosten, Zuzahlung, Eigenanteil, Mehrkosten
data4_zg3_kosten <- data4  %>% 
  group_by( V_PFLEGEGRAD,TG,V_ALTER, tod) %>%
  summarise(SUM_HM_GesKosten = sum(HM_GESKOSTEN),
            SUM_HM_Zuzahlung = sum(HM_ZUZAHLUNG),
            SUM_HM_Eigenanteil = sum(HM_EIGENANTEIL),
            SUM_HM_Mehrkosten = sum(HM_MEHRKOSTEN))

# alles verbinden
data4_zg3_a <- left_join(data1a_zg3_verweil, data4_zg3_id3, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))
data4_zg3_b <- left_join(data4_zg3_a,data4_zg3_tage2, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0)) 
data4_zg3_c <- left_join(data4_zg3_b, data4_zg3_leist2, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))
data4_zg3_d <- left_join(data4_zg3_c, data4_zg3_kosten,by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))

# Daten für Tabelle vorbereiten
results_alle_w  <- data4_zg3_d %>% 
  mutate(proz_n_IN_HM = n_IN_HM/n_ID,
         Kosten_Leistung = SUM_HM_GesKosten/Leistungen_HM,
         Leistungen_HM_jahr = Leistungen_HM*365/oKH_Tage,
         HM_GesKosten_jahr = SUM_HM_GesKosten*365/oKH_Tage)

# relevante Variablen für Tabelle
results_alle_w <- results_alle_w %>% 
  dplyr::select(TG,V_PFLEGEGRAD,V_ALTER, tod, n_ID, VZ_tage,oKH_Tage,n_IN_HM,Leistungen_HM,Kosten_Leistung, proz_n_IN_HM,Leistungen_HM_jahr,HM_GesKosten_jahr) %>% 
  mutate_all(~replace(., is.na(.), 0))  
```

```{r wichtung_2021}
# zunächst Anzahl der VZ_TAGE pro TG
data5 <- results_alle_w %>%  
  group_by(TG) %>% 
  summarise(days_tg = sum(oKH_Tage))

# mit Ergebnise verbinden 
data5a <- left_join(results_alle_w, data5, by=c("TG"))

# Wichtung nach TG1
data6 <- data5a %>% 
  filter(TG=="TG1") %>% 
  mutate(weight=oKH_Tage/days_tg) %>% 
  ungroup() %>% 
  dplyr::select(c("V_PFLEGEGRAD","V_ALTER", "tod","weight"))

# erneute Verbindung mit Ergebnissen
data6a <- left_join(data5a, data6, by=c("V_PFLEGEGRAD","V_ALTER", "tod")) %>%
  mutate_all(~replace(., is.na(.), 0))

# Wichtung anwenden
data6b <- data6a %>% mutate(Leistungen_HM_jahr_w = Leistungen_HM_jahr * weight,
                            HM_GesKosten_jahr_w = HM_GesKosten_jahr * weight)


# finale Werte
data7 <- data6b %>% 
  group_by(TG) %>% 
  filter(weight > 0) %>% 
  summarise(oKH_Tage = sum(oKH_Tage),
            Leistungen_HM_jahr = sum(Leistungen_HM_jahr_w),
            HM_GesKosten_jahr = sum(HM_GesKosten_jahr_w))

# for n
help1 <- left_join(data1,data6, by=c("V_PFLEGEGRAD","V_ALTER", "tod")) %>% 
  filter(!(is.na(weight)))

n_weight <- help1 %>% 
  group_by(TG) %>% 
  summarise(N=n_distinct(V_ID))

# zusammenführen
results_weight <- left_join(n_weight, data7, by=c("TG"))
```

```{r export_2021}
list_of_datasets <- list("results_alle" = results_alle, "results_weight" = results_weight)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/21_Kosten_SGB5_2021.xlsx")
```


# 2022
```{r data_load_2022}
setwd("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/02_rdata/2022")

load("data11_neu.Rda")
data1 <- data11_neu

load("sgbv10.Rda")
load("kh11.Rda")

verweildauer_zg3_TG <- read_excel("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2022.xlsx",3)  

verweildauer_pg_TG_tod <- read_excel("O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/13_KH_Verweildauer2022.xlsx",9) 
```


Vorbereitung KH-Daten
```{r kh_prep_2022, warning=FALSE}
### KH-Daten mit Stammdaten verbinden
# nur vollstationär behalten
kh11a<- kh11  %>%
  mutate(KH_AUF_ANLASS2 = str_pad(KH_AUF_ANLASS, 4, pad = "0"),
         KH_AUF_ANLASS_Beg = substr(KH_AUF_ANLASS2, 0, 2),
         KH_AUF_ANLASS_End = substr(KH_AUF_ANLASS2, 3, 4)) %>% 
  filter(KH_AUF_ANLASS_Beg == "01")

# überschneidende Hospitalisierungen verbinden
kh11a$KH_BEGINN_DAT <- ymd(kh11a$KH_BEGINN_DAT)
kh11a$KH_ENDE_DAT <- ymd(kh11a$KH_ENDE_DAT)
kh11a <- kh11a %>% 
  filter(!(is.na(KH_ENDE_DAT)))

kh11b<- kh11a%>%
  arrange(V_ID, KH_BEGINN_DAT) %>% 
  group_by(V_ID) %>%
  mutate(indx = c(0, cumsum(as.numeric(lead(KH_BEGINN_DAT )) > cummax(as.numeric(KH_ENDE_DAT )))[-n()])) %>%
  group_by(V_ID, indx) %>%
  summarise(KH_BEGINN_DAT2 = min(KH_BEGINN_DAT), 
            KH_ENDE_DAT2 = max(KH_ENDE_DAT))

#Klinikdaten für merge mit HM-Daten vorbereiten (rename und select)
kh11c <- kh11b %>% 
  mutate(LEISTUNG_VON = KH_BEGINN_DAT2,
         LEISTUNG_BIS = KH_ENDE_DAT2,
         origin = "KH") %>% 
  select(V_ID, LEISTUNG_VON, LEISTUNG_BIS, origin)
```


```{r, include = FALSE}
# HM-Daten vorbereiten
sgbv10$HM_DATUM <- ymd(sgbv10$HM_DATUM)
sgbv10$LEISTUNG_VON <- sgbv10$HM_DATUM
sgbv10$LEISTUNG_BIS <- sgbv10$HM_DATUM

# negative Kosten exkludieren
sgbv10 <- sgbv10 %>% filter(HM_GESKOSTEN>=0)

# HM-Datensatz für Verbindung erstellen
help1 <- sgbv10%>% arrange(V_ID) %>% 
  mutate(origin="HM")  

# Daten mit Hospitalisierungen (row_bind) und Stammdaten verbinden (filtern von fehlenden Beginn/Ende bzw. ohne Match)
help2 <- bind_rows(help1, kh11c)

data3 <- left_join(data1, help2, by = ("V_ID")) %>% 
  filter(!(is.na(LEISTUNG_VON))) %>% 
  filter(!(is.na(LEISTUNG_BIS))) 

# für jeden Tag eines Leistungserhaltes Zeile erstellen 
data3a <- data3 %>%
  # create sequence of days between start and end
  mutate(day = map2(LEISTUNG_VON, LEISTUNG_BIS, ~seq(.x, .y, "day"))) %>%
  # unnest data
  unnest(cols = c(day)) %>%                                                       
  group_by(V_ID,V_PFLEGEGRAD, TG,origin, day)

data3a2 <- data3a %>% 
  arrange(V_ID) %>% 
  ungroup() 

# nur relevante Tage  behalten (Filter vor allem notwendig wenn mehrere TGs und PG bei einer Person um passendes Intervall zu matchen, teilweise Leistungen nach Todesdatum)
data3b <- data3a2 %>%
  filter(START_DATE <= day & END_DATE >= day) %>% 
  mutate(HM = if_else(origin == "HM", 1, 0),
         kh = if_else(origin == "KH", 1, 0))

# für jede group_by Kombi prüfen ob an bestimmten Tag HM-leistung und/oder Hospitalisierung
data3c <- data3b %>% 
  group_by(V_ID, V_PFLEGEGRAD, TG, day) %>% 
  summarise(HM = max(HM),
            kh = max(kh))

# nur HM-Leistung ohne KH behalten
data3d <- data3c %>% 
  filter(!(kh == 1)) 

# HM-Leistungstage wieder mit Kosten verbinden
data3e <- left_join(data3d, sgbv10, by=c("V_ID", "day"="HM_DATUM")) %>% 
  select(-c(LEISTUNG_VON,LEISTUNG_BIS)) %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD)) 
```


Verbindung mit Gesamtdaten
```{r, include = FALSE}
data1 <- data1 %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD))

# Stammdaten mit Leistungstagen verbinden
data3f <- left_join(data1, data3e, by = c("V_ID", "V_PFLEGEGRAD", "TG"))

# nur Zeilen behalten die Stammdatenintervall und Leistungstag einschließen
data4 <- data3f %>% 
  filter(!(day > END_DATE | day < START_DATE)) %>% 
  rename(HM_DATUM = day)

data4 <- data4 %>%
  mutate(TG = as.factor(TG), 
         V_PFLEGEGRAD = as.factor(V_PFLEGEGRAD)) 

# unique Id pro Leistung
data4 <- data4 %>% mutate(id=row_number())
```


## Leistung nach TG
```{r}
# Data 1 vorbereiten
data1a_zg <- data1  %>%
  group_by(V_ID, TG) %>% 
  summarise(VZ_tage = sum(intervalldauer))

data1a_zg3 <- left_join(data1a_zg, verweildauer_zg3_TG, by=c("V_ID", "TG")) %>% 
  mutate(Verweildauer = if_else(is.na(Verweildauer),0,Verweildauer),
         oKH_Tage = VZ_tage- Verweildauer)

# Tage ohne KH_Aufenthalt
data1a_zg3_verweil <- data1a_zg3 %>% 
  group_by( TG, .drop=FALSE) %>% 
  summarise( VZ_tage = sum(VZ_tage),
             oKH_Tage = sum(oKH_Tage))


### auf ID-Basis
data4_zg3_id <- data4  %>%  
  group_by(V_ID,TG) %>% 
  summarise(HM_ja_id = 1) %>% 
  group_by(TG) %>% 
  summarise(n_IN_HM = sum(HM_ja_id))

data4_zg3_id2 <- data1  %>%  
  group_by(V_ID,TG) %>%
  summarise(id_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(n_ID = sum(id_ja))
  
data4_zg3_id3 <- left_join(data4_zg3_id2,data4_zg3_id, by=c("TG")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(n_noIN_HM = n_ID- n_IN_HM)

# Gesamttage mit HM-Leistung und Tage ohne HM-Leistung 
data4_zg3_tage1 <- data4 %>%
  group_by(V_ID,TG, HM_DATUM) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(IN_Tage_HM = sum(ID_ja))

data4_zg3_tage2<-  left_join(data1a_zg3_verweil,data4_zg3_tage1, by=c("TG")  ) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(noIN_Tage_HM= oKH_Tage -IN_Tage_HM) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtleistungen HM pro TG 
data4_zg3_leist1 <- data4 %>%  
  group_by(V_ID,TG,id) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(TG) %>% 
  summarise(Leistungen_HM = sum(ID_ja))

data4_zg3_leist2<-  left_join(data1a_zg3_verweil,data4_zg3_leist1, by=c("TG")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))


### Gesamtkosten, Zuzahlung, Eigenanteil, Mehrkosten
data4_zg3_kosten <- data4  %>% group_by(TG, .drop=FALSE) %>%
  summarise(SUM_HM_GesKosten = sum(HM_GESKOSTEN),
            SUM_HM_Zuzahlung = sum(HM_ZUZAHLUNG),
            SUM_HM_Eigenanteil = sum(HM_EIGENANTEIL),
            SUM_HM_Mehrkosten = sum(HM_MEHRKOSTEN))


# alles verbinden
data4_zg3_a <- left_join(data1a_zg3_verweil, data4_zg3_id3, by=c("TG"))
data4_zg3_b <- left_join(data4_zg3_a,data4_zg3_tage2, by=c("TG"))
data4_zg3_c <- left_join(data4_zg3_b, data4_zg3_leist2, by=c("TG")) 
data4_zg3_d <- left_join(data4_zg3_c, data4_zg3_kosten, by=c("TG")) 

# Daten für Tabelle vorbereiten
results_alle <- data4_zg3_d %>% 
  mutate(proz_n_IN_HM = n_IN_HM/n_ID,
         Kosten_Leistung = SUM_HM_GesKosten/Leistungen_HM,
         Leistungen_HM_jahr = Leistungen_HM*365/oKH_Tage,
         HM_GesKosten_jahr = SUM_HM_GesKosten*365/oKH_Tage)

# relevante Variablen für Tabelle
results_alle <- results_alle %>% 
  dplyr::select(TG, n_ID, VZ_tage,oKH_Tage,n_IN_HM,Leistungen_HM,
                proz_n_IN_HM,Kosten_Leistung,Leistungen_HM_jahr,HM_GesKosten_jahr) %>% 
  mutate_all(~replace(., is.na(.), 0))  
```

## Leistung nach ZG, PG und TG
```{r}
# Data 1 vorbereiten
data1a_zg <- data1  %>%  
  group_by(V_ID, V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(VZ_tage = sum(intervalldauer ))

verweildauer_pg_TG_tod$V_ALTER <- as.integer(verweildauer_pg_TG_tod$V_ALTER )
data1a_zg3 <- left_join(data1a_zg, verweildauer_pg_TG_tod, by=c("V_ID", "V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate(Verweildauer = if_else(is.na(Verweildauer), 0, Verweildauer),
         oKH_Tage = VZ_tage - Verweildauer )

# Tage ohne KH_Aufenthalt 
data1a_zg3_verweil <- data1a_zg3 %>% 
  group_by( V_PFLEGEGRAD,TG,V_ALTER, tod, .drop=FALSE) %>% 
  summarise(VZ_tage = sum(VZ_tage),
            oKH_Tage = sum(oKH_Tage))

# auf ID-Basis 
data4_zg3_id <- data4  %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(HM_ja_id = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(n_IN_HM = sum(HM_ja_id))

data4_zg3_id2 <- data1  %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod) %>%
  summarise(id_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(n_ID= sum(id_ja))
  
data4_zg3_id3 <- left_join(data4_zg3_id2,data4_zg3_id, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod") ) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(n_noIN_HM=n_ID- n_IN_HM)

# Gesamttage mit HM-Leistung und Tage ohne HM-Leistung 
data4_zg3_tage1 <- data4 %>%  
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod, HM_DATUM) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(IN_Tage_HM = sum(ID_ja) )

data4_zg3_tage2<-  left_join(data1a_zg3_verweil,data4_zg3_tage1,by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  mutate(noIN_Tage_HM = oKH_Tage -IN_Tage_HM) %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtleistungen HM pro TG 
data4_zg3_leist1 <- data4 %>%
  group_by(V_ID,V_PFLEGEGRAD,TG,V_ALTER, tod,id) %>% 
  summarise(ID_ja = 1) %>% 
  group_by(V_PFLEGEGRAD,TG,V_ALTER, tod) %>% 
  summarise(Leistungen_HM = sum(ID_ja))

data4_zg3_leist2<-  left_join(data1a_zg3_verweil,data4_zg3_leist1, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0))  %>% 
  dplyr::select(-c(VZ_tage,oKH_Tage))

# Gesamtkosten, Zuzahlung, Eigenanteil, Mehrkosten
data4_zg3_kosten <- data4  %>% 
  group_by( V_PFLEGEGRAD,TG,V_ALTER, tod) %>%
  summarise(SUM_HM_GesKosten = sum(HM_GESKOSTEN),
            SUM_HM_Zuzahlung = sum(HM_ZUZAHLUNG),
            SUM_HM_Eigenanteil = sum(HM_EIGENANTEIL),
            SUM_HM_Mehrkosten = sum(HM_MEHRKOSTEN))

# alles verbinden
data4_zg3_a <- left_join(data1a_zg3_verweil, data4_zg3_id3, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))
data4_zg3_b <- left_join(data4_zg3_a,data4_zg3_tage2, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod")) %>% 
  mutate_all(~replace(., is.na(.), 0)) 
data4_zg3_c <- left_join(data4_zg3_b, data4_zg3_leist2, by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))
data4_zg3_d <- left_join(data4_zg3_c, data4_zg3_kosten,by=c("V_PFLEGEGRAD", "TG","V_ALTER", "tod"))

# Daten für Tabelle vorbereiten
results_alle_w  <- data4_zg3_d %>% 
  mutate(proz_n_IN_HM = n_IN_HM/n_ID,
         Kosten_Leistung = SUM_HM_GesKosten/Leistungen_HM,
         Leistungen_HM_jahr = Leistungen_HM*365/oKH_Tage,
         HM_GesKosten_jahr = SUM_HM_GesKosten*365/oKH_Tage)

# relevante Variablen für Tabelle
results_alle_w <- results_alle_w %>% 
  dplyr::select(TG,V_PFLEGEGRAD,V_ALTER, tod, n_ID, VZ_tage,oKH_Tage,n_IN_HM,Leistungen_HM,Kosten_Leistung, proz_n_IN_HM,Leistungen_HM_jahr,HM_GesKosten_jahr) %>% 
  mutate_all(~replace(., is.na(.), 0))  
```

```{r wichtung_2022}
# zunächst Anzahl der VZ_TAGE pro TG
data5 <- results_alle_w %>%  
  group_by(TG) %>% 
  summarise(days_tg = sum(oKH_Tage))

# mit Ergebnise verbinden 
data5a <- left_join(results_alle_w, data5, by=c("TG"))

# Wichtung nach TG1
data6 <- data5a %>% 
  filter(TG=="TG1") %>% 
  mutate(weight=oKH_Tage/days_tg) %>% 
  ungroup() %>% 
  dplyr::select(c("V_PFLEGEGRAD","V_ALTER", "tod","weight"))

# erneute Verbindung mit Ergebnissen
data6a <- left_join(data5a, data6, by=c("V_PFLEGEGRAD","V_ALTER", "tod")) %>%
  mutate_all(~replace(., is.na(.), 0))

# Wichtung anwenden
data6b <- data6a %>% mutate(Leistungen_HM_jahr_w = Leistungen_HM_jahr * weight,
                            HM_GesKosten_jahr_w = HM_GesKosten_jahr * weight)


# finale Werte
data7 <- data6b %>% 
  group_by(TG) %>% 
  filter(weight > 0) %>% 
  summarise(oKH_Tage = sum(oKH_Tage),
            Leistungen_HM_jahr = sum(Leistungen_HM_jahr_w),
            HM_GesKosten_jahr = sum(HM_GesKosten_jahr_w))

# for n
help1 <- left_join(data1,data6, by=c("V_PFLEGEGRAD","V_ALTER", "tod")) %>% 
  filter(!(is.na(weight)))

n_weight <- help1 %>% 
  group_by(TG) %>% 
  summarise(N=n_distinct(V_ID))

# zusammenführen
results_weight <- left_join(n_weight, data7, by=c("TG"))
```

```{r export_2022}
list_of_datasets <- list("results_alle" = results_alle, "results_weight" = results_weight)

write.xlsx(list_of_datasets, file = "O:/U5279/Routinedaten/Abschlussdatenlieferung_neu_202302/Daten/03_Ergebnisse/21_Kosten_SGB5_2022.xlsx")
```
